% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_class}
\alias{add_class}
\title{Add a class to a R object}
\usage{
add_class(x, class_name, replace = FALSE)
}
\arguments{
\item{x}{a R object}

\item{class_name}{the class name to be assigned to x}

\item{replace}{logical. Whether the class_name should be added to (FALSE,
default) or should replace (TRUE) existing classes.}
}
\value{
the R object \code{x} with updated class assignment
}
\description{
Simple wrapper around the {base} \code{class} function to allow its use in
{magrittr} pipe workflow
}
