% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{alliance_finish}
\alias{alliance_finish}
\title{Alliance Finish}
\usage{
alliance_finish(status)
}
\arguments{
\item{status}{status object as output in a column of event_alliances()}
}
\description{
Switching fxn to use appropriate "alliance_finish" status (double elim or
bo3) based on the internal structure of the input status object
}
