% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_indexer}
\alias{apply_indexer}
\title{Apply Indexer}
\usage{
apply_indexer(df, idx)
}
\description{
Applies the indexer function to the dataframe and pulls out the diagonal
of the resulting matrix, efficiently getting the requested data from DF.
}
\details{
Assumes that the 'df' parameter is coercable to a matrix. Also
assumes that idx is a dataframe with columns "ridx" and "cidx" (standing for
row index and col index) that index into df.
}
\author{
Isolde Moyer
}
