% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{attach_opt_params}
\alias{attach_opt_params}
\title{Attach Optional Parameters}
\usage{
attach_opt_params(req, params, marker = "?")
}
\arguments{
\item{req}{Length-1 character vector representing an API request}

\item{params}{named list of desired params, where the name is the parameter
title and the value is the parameter value.}

\item{marker}{the character used to separate the params in the url}
}
\description{
Attaches optional parameters to an API request
}
