% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{attach_parameters}
\alias{attach_parameters}
\title{Attach parameters}
\usage{
attach_parameters(req, parameters, values)
}
\arguments{
\item{req}{Request string prior to operation}

\item{parameters}{Vector of parameter titles}

\item{values}{Vector of parameter values}
}
\value{
String request with parameter attached
}
\description{
Helper function to write API calls, adds parameters to the request
}
\examples{
attach_parameters("api.example/v2/root", "option", "2")
}
\author{
Gabriel Krotkov
}
