% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{auth}
\alias{auth}
\title{API Authorization}
\usage{
auth(req)
}
\arguments{
\item{req}{Request string stub}
}
\value{
string request with TBA auth key
}
\description{
Attaches the auth key to the end of a request string
}
\examples{
auth("")
}
\author{
Gabriel Krotkov
}
