% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{awards24}
\alias{awards24}
\title{All Team Awards}
\format{
An object of class \code{list} of length 8410.
}
\source{
thebluealliance.com
}
\usage{
awards24
}
\description{
List of all team official award histories as of the 2024 season
}
\keyword{datasets}
