% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{bo3_alliance_finish}
\alias{bo3_alliance_finish}
\title{Best of 3 Bracket Alliance Finish}
\usage{
bo3_alliance_finish(status)
}
\arguments{
\item{status}{alliance "status" object as return in the column of the output
of scoutR::event_alliances(). A list, usually of length 8.}
}
\description{
Returns a character vector representing the finish of the alliances from a
single elimation, best of three bracket.
}
\details{
case_when uses a "step-down" resolution; so we rely on the first condition
being false to make later conditions correctly specified.
}
