% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{breakdown}
\alias{breakdown}
\title{Retrieve score breakdown object}
\usage{
breakdown(match)
}
\arguments{
\item{match}{TBA match object}
}
\value{
Score breakdown object (list)
}
\description{
Returns a score breakdown object given a match object
}
\examples{
paphi <- read_event_matches("2016paphi")
breakdown(paphi[[1]])
}
\author{
Gabriel Krotkov
}
