% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMAD.R
\name{calcMAD}
\alias{calcMAD}
\title{calcMAD}
\usage{
calcMAD(x, variable, nmads, bound = c("higher", "lower"))
}
\arguments{
\item{x}{A Seurat object}

\item{variable}{A valid column name in metadata to be tested}

\item{nmads}{The number of median absolute deviations to consider}

\item{bound}{One of "higher", or "lower"}
}
\value{
A single numeric value representing the value at the specified nMAD
}
\description{
Calculates median average deviation for a set value for a variable of
interest to help quantitatively choose thresholds
}
\examples{
# Calculate an upper bound on a single Seurat object
calcMAD(
  x = seurat_obj,
  variable = "log_atac_peak_region_fragments",
  nmads = 5,
  bound = "higher")

}
