% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_first}
\alias{cap_first}
\title{Capitalize First Character}
\usage{
cap_first(s)
}
\arguments{
\item{s}{input character vector}
}
\description{
Capitalizes the first character of each of the elements of the input vector
}
