% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_tbls}
\alias{combine_tbls}
\title{Combine Tables}
\usage{
combine_tbls(tbl1, tbl2)
}
\arguments{
\item{tbl1}{First table to be combined}

\item{tbl2}{Second table to be combined}
}
\description{
Combines two tables, numerical vectors of integer counts. Assumes that
order() is a meaningful call on the names of the tables.
}
