% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{count_team_matches}
\alias{count_team_matches}
\title{Count Team Matches}
\usage{
count_team_matches(matches)
}
\arguments{
\item{matches}{dataframe assumed to have "blue1", "blue2", "blue3", "red1",
"red2", and "red3" columns.}
}
\value{
dataframe with team ids and the count of matches appeared in
}
\description{
Counts the number of times a team appears in a matches dataframe
}
