% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{de_alliance_finish}
\alias{de_alliance_finish}
\title{Double Elimination Alliance Finish}
\usage{
de_alliance_finish(status)
}
\arguments{
\item{status}{alliance "status" object as returned in a column of the output
of event_alliances(). A list, usually of length 8.}
}
\description{
Returns a character vector representing the finish (placement) of the
alliances from a double elimination bracket.
}
\details{
case_when uses a "step-down" resolution; so we rely on the first condition
being false to make later conditions correctly specified.
}
