% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{district_events}
\alias{district_events}
\title{District Events}
\usage{
district_events(district_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{district_key}{TBA legal district key (remember year as prefix)}

\item{simple}{(bool) simplify event objects?}

\item{keys}{(bool) return keys only?}
}
\value{
tidy tibble of district events, or vector if keys = TRUE
}
\description{
Reads events in a given district
}
\examples{
district_events("2016mar")
district_events("2018fma", keys = TRUE)
}
