% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{district_rankings}
\alias{district_rankings}
\title{District Rankings}
\usage{
district_rankings(
  district_key,
  separate_events = FALSE,
  event_breakdown = FALSE
)
}
\arguments{
\item{district_key}{(character) TBA legal district key}

\item{separate_events}{(bool) split up team performance across each event?}

\item{event_breakdown}{(bool) breakdown event performance?}
}
\value{
tidy tibble of event rankings
}
\description{
Reads district rankings
}
\examples{
district_rankings("2016mar", separate_events = TRUE, event_breakdown = TRUE)
district_rankings("2022fit")
}
\author{
Gabriel Krotkov
}
