% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{district_teams}
\alias{district_teams}
\title{District Teams}
\usage{
district_teams(district_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{district_key}{TBA legal district key}

\item{simple}{(bool) simplify team objects?}

\item{keys}{(bool) get keys only?}
}
\value{
Tidy tibble of teams objects in a given district (or vector if keys)
}
\description{
Gets teams in a given district-year
}
\examples{
district_teams("2022fit", keys = TRUE)
district_teams("2014mar", simple = TRUE)
district_teams("2016mar")
}
\author{
Gabriel Krotkov
}
