% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{districts}
\alias{districts}
\title{Districts}
\usage{
districts(year = YEAR)
}
\arguments{
\item{year}{(int) year of interest}
}
\value{
tidy tibble of districts in the given year
}
\description{
Reads districts in a given year
}
\examples{
districts(2012)
districts(2023)
}
\author{
Gabriel Krotkov
}
