% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{document}
\alias{document}
\title{Document}
\usage{
document()
}
\description{
Automatically creates .rd documentation files for each function in scoutR
}
\details{
Uses here() to always call from the project root.
}
