% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dquals24}
\alias{dquals24}
\title{District Qualification Data}
\format{
An object of class \code{list} of length 827.
}
\source{
thebluealliance.com
}
\usage{
dquals24
}
\description{
A list of qualification data from every district played between 2009 and
2024. Every item in the list is a dataframe, and each dataframe has at least
the columns "red1", "red2", "red3", "blue1", "blue2", "blue3", "red_score",
and "blue_score". Generated by scripts/get_district_quals.R
}
\keyword{datasets}
