% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{dual_line_weight_approximation}
\alias{dual_line_weight_approximation}
\title{Dual Line Weight Approximation}
\usage{
dual_line_weight_approximation(bin_vars)
}
\arguments{
\item{bin_vars}{A vector of the variance in each bin}
}
\value{
vector of weights
}
\description{
Given the variance in a number of bins, approximate an appropriate WLS
weighting by two lines.
}
