% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_possession_sequence.R
\name{encode_possession_sequence}
\alias{encode_possession_sequence}
\title{Encode single-match possessions}
\usage{
encode_possession_sequence(data, event_var, team_var, match = NULL)
}
\arguments{
\item{data}{A data frame of event data.}

\item{event_var}{A character value identifying the variable name that
corresponds to the event classification for a given event in \code{data}.}

\item{team_var}{A character value identifying the variable name that
corresponds to the team label for a given event in \code{data}.}

\item{match}{A character vector of match id values.
Only required to define possession ids and sequences in
multi-match data frames. See example below.}
}
\value{
A tidy tibble that contains \code{data}, and \code{possession_id} and
\code{possession_seq} variables. Variable \code{possession_id} maps each
event to a possession within a game. Variable \code{possession_seq}
provides a running count of events within a possession.
}
\description{
This function encodes possession sequences for a single match using event
data.
}
\examples{
# load dplyr and purrr for examples
library(dplyr)
library(purrr)

# read event and teams data given in package
events <- system.file("extdata", "events_england.json", package = "scoutr") \%>\%
  read_events()
teams <- system.file("extdata", "teams.json", package = "scoutr") \%>\%
  read_teams()

# join events and teams data
x <- left_join(events, teams, by = c("team_id" = "wy_id")) \%>\%
  select(match_id, event_name, sub_event_name, team_id, name)

encode_possession_sequence(x, event_var = "event_name", team_var = "name")

# multi-match example (fake)
y <- left_join(events, teams, by = c("team_id" = "wy_id")) \%>\%
  select(match_id, event_name, sub_event_name, team_id, name) \%>\%
  slice(1:10) \%>\%
  mutate(match_id = as.character(c(rep(1, 5), rep(2, 5))))

map_df(unique(y$match_id), encode_possession_sequence, data = y,
                           event_var = "event_name", team_var = "name")

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
