% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event}
\alias{event}
\title{Get Event Object}
\usage{
event(key, simple = FALSE)
}
\arguments{
\item{key}{TBA legal event key}

\item{simple}{(bool) simplify event objects?}
}
\value{
tidy tibble of one row with an event object
}
\description{
Reads event object from TBA API
}
\examples{
event("2022txirv")
event("2016pahat", simple = TRUE)
}
\author{
Gabriel Krotkov
}
