% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_alliances}
\alias{event_alliances}
\title{Event Alliances}
\usage{
event_alliances(key, unpack_picks = TRUE, attach_finish = TRUE)
}
\arguments{
\item{key}{TBA legal event key}

\item{unpack_picks}{(logical) break out alliance column?}

\item{attach_finish}{(logical)}
}
\value{
Tidy tibble of event alliances
}
\description{
Read event alliances
}
\examples{
event_alliances("2016hop", unpack_picks = TRUE)
event_alliances("2015paphi")
event_alliances("2014mrcmp", unpack_picks = FALSE)
}
\author{
Gabriel Krotkov
}
