% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_awards}
\alias{event_awards}
\title{Event Awards}
\usage{
event_awards(key, recipients = FALSE)
}
\arguments{
\item{key}{TBA legal event key}

\item{recipients}{(bool) unpack recipients column?}
}
\value{
tidy tibble of event awards
}
\description{
Reads event awards
}
\examples{
event_awards("2015padre")
event_awards("2010ein", recipients = TRUE)
}
\author{
Gabriel Krotkov
}
