% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_coprs}
\alias{event_coprs}
\title{Get Event cOPRs}
\usage{
event_coprs(key, digits = 2)
}
\arguments{
\item{key}{(string) event key}

\item{digits}{number of digits to round the result to}
}
\description{
Wrapper function for read_event_coprs and tidy_event_coprs
}
\details{
TBA exports cOPRs for events as early as 2023. If you want cOPRs for earlier
events, you can compute them with `fit_event_lr` and a wise response choice
}
