% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_district_points}
\alias{event_district_points}
\title{Event District Points}
\usage{
event_district_points(key, tiebreakers = FALSE)
}
\arguments{
\item{key}{TBA legal event key}

\item{tiebreakers}{(bool) unpack tiebreakers?}
}
\value{
tidy tibble of district points?
}
\description{
Reads event district points
}
\examples{
event_district_points("2016mrcmp", tiebreakers = TRUE)
event_district_points("2015padre")
}
\author{
Gabriel Krotkov
}
