% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_insights}
\alias{event_insights}
\title{Event Insights}
\usage{
event_insights(key, insight_as_col = FALSE)
}
\arguments{
\item{key}{event key}

\item{insight_as_col}{(bool) Will we treat the "phase" of insights
(qual/playoff) as the column variable? If TRUE, will return a dataframe
with the variables "field", "qual", "playoff", (and any other phases.) If
FALSE, will have each field in its own column and a row for each phase.}
}
\description{
Read event insights
}
\details{
We assume that sublists have the same number and names for fields.
That is, quals and playoffs will have the same length of list, and have
the same fields filled out in both lists, in the same order.
}
