% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_matches}
\alias{event_matches}
\title{Event Matches}
\usage{
event_matches(
  event_key,
  alliances = TRUE,
  breakdown = TRUE,
  match_type = "all",
  unplayed = FALSE,
  simple = FALSE,
  keys = FALSE
)
}
\arguments{
\item{event_key}{TBA legal event key}

\item{alliances}{(bool) break down alliances column?}

\item{breakdown}{(bool) break down score_breakdown column?}

\item{match_type}{one of "all", "qual", or "playoff"}

\item{unplayed}{(boolean) include matches with scores of -1 (indicating that the match has not been played?)}

\item{simple}{(bool) simplify match object?}

\item{keys}{(bool) get keys only?}
}
\value{
Tibble of matches from the given event.
}
\description{
Reads matches of a given event
}
\examples{
event_matches("2015paphi", match_type = "playoff")
event_matches("2014mrcmp", match_type = "qual", keys = TRUE)
}
\author{
Gabriel Krotkov
}
