% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{event_matchups}
\alias{event_matchups}
\title{Event Matchups}
\usage{
event_matchups(event_code, team_id)
}
\arguments{
\item{event_code}{TBA-legal event code (ex. "2025chcmp")}

\item{team_id}{team of interest as int or TBA id (ex. "frc449", 449)}
}
\description{
Returns a list containing all a team's partners and opponents for a
particular event.
}
