% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_oprs}
\alias{event_oprs}
\title{Event OPRs}
\usage{
event_oprs(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
tidy tibble of event OPRs with columns: opr, dpr, ccwm, team
}
\description{
Read event OPRs
}
\examples{
event_oprs("2015paphi")
}
\author{
Gabriel Krotkov
}
