% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_rankings}
\alias{event_rankings}
\title{Event Rankings}
\usage{
event_rankings(key, trim = TRUE)
}
\arguments{
\item{key}{TBA legal event key}

\item{trim}{(bool) remove data unrelated to game analysis?}
}
\value{
tidy tibble of event rankings
}
\description{
Read event rankings
}
\examples{
event_rankings("2016hop")
}
\author{
Gabriel Krotkov
}
