% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{event_robot_results}
\alias{event_robot_results}
\title{Robot Results}
\usage{
event_robot_results(event_code, match_type = "all")
}
\arguments{
\item{event_code}{Event code of interest}

\item{match_type}{One of "all", "qual", or "playoff"}
}
\description{
Get all the robot-level results available in TBA, under a few assumptions.
}
\details{
Assumes that names of robot-level information follow the convention:
"(red/blue)_robot_(1/2/3)"
}
\examples{
mil23_individual <- event_robot_results("2023mil", match_type = "qual")
gpr24_individual <- event_robot_results("2024paca")

}
