% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{event_sb}
\alias{event_sb}
\title{Event (Statbotics)}
\usage{
event_sb(key)
}
\arguments{
\item{key}{event key (e.g. "2025mdpas")}
}
\description{
Returns a 1-row dataframe for a given event
}
\examples{
pasadena <- event_sb("2025mdpas")
}
