% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{event_season_history}
\alias{event_season_history}
\title{Event Season History}
\usage{
event_season_history(event_code, fields = NULL, manual_teams = NULL)
}
\arguments{
\item{event_code}{TBA-legal event code (ex. "2024paca")}

\item{fields}{optional, robot fields you want to retrieve. If NULL, uses
`id_robot_fields()` to get individual robot-level fields automatically.}

\item{manual_teams}{(int) additional teams to manually add to the team list}
}
\description{
Given an event code, this function returns all a dataframe with all the
robot-level results of teams registered for a particular event.
}
\details{
Checks for match duplication, which will stop execution if TRUE.
}
\examples{
gpr24 <- event_season_history("2024paca")
}
