% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{event_tangibles}
\alias{event_tangibles}
\title{Event Tangibles}
\usage{
event_tangibles(
  event_key,
  schema = schema_cfs,
  qual_only = T,
  pct = T,
  digits = 2
)
}
\arguments{
\item{event_key}{TBA-legal event key (i.e. "2025vagle")}

\item{schema}{function defining schema for column names. schema_cfs() has
been the standard for 2018 - 2025 (and likely beyond)}

\item{qual_only}{(logical) include only qual matches? If FALSE, will include
both qualification and playoff matches.}

\item{pct}{Convert all data columns to a percentage rather than a count?}

\item{digits}{if non-NULL, rounds all numeric columns to the number of digits
supplied.}
}
\description{
Pulls the "tangible" results for a given event from TBA. Identifies
"tangible" fields by the standard "robot" naming schema that TBA and the
FIRST API tends to use to describe any robot-specific fields.
}
\examples{
event_tangibles("2025vagle")
event_tangibles("2025vagle", qual_only = FALSE)
event_tangibles("2017mrcmp", schema = schema_csf)
}
