% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_teams}
\alias{event_teams}
\title{Event Teams}
\usage{
event_teams(key, statuses = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{key}{TBA legal event key}

\item{statuses}{(bool) unpack event statuses?}

\item{simple}{(bool) simplify teams objects?}

\item{keys}{(bool) get keys only?}
}
\description{
Reads teams at a given event
}
\examples{
event_teams("2015padre", keys = TRUE)
}
\author{
Gabriel Krotkov
}
