% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{event_winner_seed}
\alias{event_winner_seed}
\title{Event Winner Seed}
\usage{
event_winner_seed(key)
}
\arguments{
\item{key}{TBA-legal event key (ex. "2025vagle")}
}
\description{
Given an event, retrieves the seed of the winning alliance.
}
\examples{
event_winner_seed("2025vagle")

}
