% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{events}
\alias{events}
\title{Events}
\usage{
events(year, official = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{year}{year of interest}

\item{official}{(bool) get only official events?}

\item{simple}{(bool) simplify event objects?}

\item{keys}{(bool) get keys only?}
}
\value{
tidy tibble of event objects, or vector if keys = TRUE
}
\description{
Reads a group of event objects from TBA API
}
\examples{
events(2019)
events(2010, official = TRUE)
events(2012, simple = TRUE)
events(2015, keys = TRUE)
}
\author{
Gabriel Krotkov
}
