% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{events_sb}
\alias{events_sb}
\title{Events (Statbotics)}
\usage{
events_sb(...)
}
\arguments{
\item{...}{Parameters passed to the statbotics /v3/{events} endpoint}
}
\description{
Returns a batch of statbotics event objects conforming to the given parameters
}
