% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_read_competitions.R
\name{fc_read_competitions}
\alias{fc_read_competitions}
\title{Read Wyscout soccer competitions data}
\usage{
fc_read_competitions(file, tidy_tibble = TRUE)
}
\arguments{
\item{file}{File path to \code{competitions} JSON data.}

\item{tidy_tibble}{Logical value indicating if the returned object should be
a tidy tibble or not. If \code{FALSE}, a list is returned.}
}
\value{
A tidy tibble or list depending on the value of argument
\code{tidy_tibble}. A detailed description of the variables is given
at the reference cited below.
}
\description{
This function reads and processes the Wyscout \code{competitions} data.
}
\note{
The variable names in the returned tidy tibble are in snake case, but
they are analogous to those used by Wyscout.
}
\examples{
file_path <- system.file("extdata", "competitions.json", package = "scoutr")

competitions <- fc_read_competitions(file_path)
competitions_list <- fc_read_competitions(file_path, tidy_tibble = FALSE)

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
\seealso{


Other read functions: 
\code{\link{fc_read_coaches}()},
\code{\link{fc_read_events}()},
\code{\link{fc_read_matches}()},
\code{\link{fc_read_players}()},
\code{\link{fc_read_teams}()}
}
\concept{read functions}
