% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_defaults.r
\name{fc_theme}
\alias{fc_theme_bw}
\alias{fc_theme_classic}
\alias{fc_theme_dark}
\alias{fc_theme_gw}
\alias{fc_theme_smurf}
\alias{fc_theme_wc}
\title{Pitch themes}
\usage{
fc_theme_bw()

fc_theme_classic()

fc_theme_dark()

fc_theme_gw()

fc_theme_smurf()

fc_theme_wc()
}
\description{
These are template themes that can be matched (or not) with the
\code{palatte} argument in \code{fc_annotate_pitch()}. Use
\code{ggplot2} \code{theme()} for further theme refinements.
}
\details{
\describe{

\item{\code{fc_theme_bw()}}{
A white pitch with black field markings.}

\item{\code{fc_theme_classic()}}{
A grass-green pitch with white field markings.}

\item{\code{fc_theme_dark()}}{
A black pitch with white field markings.}

\item{\code{fc_theme_gw()}}{
A white pitch with grey field markings.}

\item{\code{fc_theme_smurf()}}{
A blue pitch with white field markings. Inspired by the Boise State
football field in Idaho.}

\item{\code{fc_theme_wc()}}{
Pitch and field markings representing the upcoming World Cup host's flag.
The 2022 World Cup is hosted by Qatar.}
}
}
\examples{
# load ggplot2
library(ggplot2)

ggplot() +
  fc_annotate_pitch() +
  fc_theme_gw()

ggplot() +
  fc_annotate_pitch(palette = "smurf") +
  fc_theme_smurf()

ggplot() +
  fc_annotate_pitch() +
  fc_theme_wc()

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
