% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_velocity_event.R
\name{fc_velocity_event}
\alias{fc_velocity_event}
\title{Compute velocity of events}
\usage{
fc_velocity_event(data, start_loc, end_loc, direction = "all")
}
\arguments{
\item{data}{A data frame with event location data. \code{data} at minimum
must contain variables named \code{match_id}, \code{match_period},
\code{possession_id}, and \code{event_sec}.}

\item{start_loc}{A character vector identifying the \code{x} and \code{y}
coordinates for the start of an event in \code{data}.}

\item{end_loc}{A character vector identifying the \code{x} and \code{y}
coordinates for the end of an event in \code{data}.}

\item{direction}{A character vector specifying the direction of velocity to
compute. It can be a subset of \code{c("east", "west", "north", "south",
  "east_west", "north_south")}. The default value computes all six velocities
and the total \code{speed}. Setting \code{direction = NULL} will only
compute the total \code{speed}.}
}
\value{
A tidy tibble with mutated columns giving the velocities specified
by \code{direction}.
}
\description{
This function computes velocities for each event in the event location
data.
}
\examples{
# load dplyr for examples
library(dplyr)

# read event data given in package
file_path <- system.file("extdata", "events_england.json", package = "scoutr")

events <- fc_read_events(file_path) \%>\%
  fc_locations_transform(x = c("start_x", "end_x"), y = c("start_y", "end_y")) \%>\%
  fc_sequence_possession(event_var = "event_name", team_var = "team_id") \%>\%
  select(match_id, match_period, possession_id, event_sec:end_y)

# compute all velocities and speed
events \%>\%
  fc_velocity_event(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y"),
                    direction = "all")

# compute speed
events \%>\%
  fc_velocity_event(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y"),
                    direction = NULL)

# compute a subset of velocities
events \%>\%
  fc_velocity_event(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y"),
                    direction = c("east_west", "north_south"))

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
