% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_velocity_polygon.R
\name{fc_velocity_polygon}
\alias{fc_velocity_polygon}
\title{Compute velocity within a grid of polygons}
\usage{
fc_velocity_polygon(
  data,
  metric,
  match = NULL,
  fcn = "median",
  ...,
  shape = "square",
  size = 5,
  preview_grid = FALSE
)
}
\arguments{
\item{data}{A simple feature object. Event velocities should be variables in
\code{data} as computed from function \code{fc_velocity_event()}.}

\item{metric}{A character vector of length one indicating which polygon
velocity to compute. Should be one of "speed", "east_velocity",
"west_velocity", "north_velocity", "south_velocity", "east_west_velocity",
"north_south_velocity".}

\item{match}{A character vector of match id values. Only required to define
possession ids and sequences in multi-match data frames.}

\item{fcn}{A summary function as a character vector. Typical values are
"mean", "median", "sd", "IQR".}

\item{...}{Arguments and their values related to \code{fcn}.}

\item{shape}{A character vector of length one indicating the polygon type.
Valid values include "square" and "hexagon".}

\item{size}{A numeric vector of length one defining the size of the polygons}

\item{preview_grid}{A logical vector of length one. A value of \code{TRUE} will
show the partitioned pitch according to the specified polygons.}
}
\value{
A simple feature object with a \code{geometry} and summary statistic
as defined by arguments \code{fcn} and \code{metric}.
}
\description{
This function computes summary statistics of velocities within a grid of
polygons that cover the pitch dimensions
}
\examples{
# load dplyr for examples
library(dplyr)

file_path <- system.file("extdata", "events_england.json", package = "scoutr")
events <- fc_read_events(file_path)

data <- events \%>\%
  select(-starts_with("tag_")) \%>\%
  fc_locations_transform(x = c("start_x", "end_x"), y = c("start_y", "end_y")) \%>\%
  fc_sequence_possession(event_var = "event_name", team_var = "team_id") \%>\%
  fc_velocity_event(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y")) \%>\%
  fc_locations_link(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y"))


fc_velocity_polygon(data, metric = "east_west_velocity", shape = "hexagon",
                    fcn = "median", na.rm = TRUE, size = 6,
                    preview_grid = TRUE)

fc_velocity_polygon(data, metric = "speed", size = 5, shape = "square",
                    fcn = "mean", na.rm = TRUE, preview_grid = TRUE)

fc_velocity_polygon(data, metric = "north_south_velocity",
                    size = 5, shape = "square",
                    fcn = "median", na.rm = TRUE, preview_grid = FALSE)

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
