% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{fit_lineup_lm}
\alias{fit_lineup_lm}
\title{Fit Lineup Linear Model}
\usage{
fit_lineup_lm(lineups, responses, w = NULL)
}
\arguments{
\item{lineups}{A dataframe of lineups assumed to have the columns `blue1`,
`blue2`, `blue3`, `red1`, `red2`, and `red3`. This can be like a dataframe
of matches as output by `event_matches`.}

\item{responses}{A list of two vectors, `red` and `blue`. Each must be a
vector of the same length as the number of rows in `lineups`, representing
the response value to fit a linear model to.}

\item{w}{Weights for WLS fit. Weights will be extended with `rep_len` to have
length corresponding to the rows of `lineups`.}
}
\description{
Computes the lineup design matrix for a given dataframe `lineups`, and then
fits a linear model using the given `responses`.
}
\details{
Key assumption - the order of `lineups` and the `responses` vector
must line up exactly. Otherwise, the fit will be meaningless.
}
\examples{
gpr <- event_matches("2024paca", match_type = "qual")
opr <- fit_lineup_lm(gpr, list(red = gpr$red_score, blue = gpr$blue_score))
}
