% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{get_content}
\alias{get_content}
\title{Get Content}
\usage{
get_content(req, base = TBA_BASE)
}
\arguments{
\item{req}{request string}

\item{base}{API base string (for example, TBA or Statbotics API root)}
}
\value{
JSON list of result string
}
\description{
Wrapper for get_reponse, uses content() to retrieve a list of the content
}
\examples{
get_content("team/frc1712/awards")
get_content("team/3504", base = STATBOTICS_BASE)
}
\author{
Gabriel Krotkov
}
