% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{get_multifield_df}
\alias{get_multifield_df}
\title{Get Dataframe for multiple fields}
\usage{
get_multifield_df(
  matches,
  fields = NULL,
  schema = schema_cfs,
  unlist = TRUE,
  merge = TRUE
)
}
\arguments{
\item{matches}{dataframe with matches on the rows}

\item{fields}{vector of name of the fields you want pulled out. If NULL,
scoutR will guess with id_robot_fields()}

\item{schema}{function defining schema for column names}

\item{unlist}{(boolean) unlist the result? Vast majority of time TRUE, FALSE
if the content of a given field has complicated content not fit for a vector}

\item{merge}{if TRUE, simplifying the resulting list of dataframes to a
single dataframe using Reduce}
}
\description{
Applies the get_robot_field_df function to all fields provided and returns a
dataframe with all the resulting data
}
\details{
This assumes that all the fields considered are categorical - it
will treat numeric variables as categorical.
}
\examples{
mil23 <- event_matches("2023mil")
fields <- id_robot_fields(mil23)
get_multifield_df(mil23, fields)
fma17 <- event_matches("2017mrcmp")
fields <- c("auto")
get_multifield_df(fma17, fields, schema = schema_csf)
}
