% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{get_response}
\alias{get_response}
\title{Get Response}
\usage{
get_response(req, base = TBA_BASE)
}
\arguments{
\item{req}{request string}

\item{base}{API base string (for example, TBA or Statbotics API root)}
}
\value{
API response
}
\description{
Wrapper for httr:GET attaches TBA base and auth key to input request
}
\examples{
get_response("team/frc1712/awards")
get_response("team/3504", base = STATBOTICS_BASE)
}
\author{
Gabriel Krotkov
}
