% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{get_robot_field_df}
\alias{get_robot_field_df}
\title{Get Field Dataframe (Robot-level)}
\usage{
get_robot_field_df(matches, field, schema = schema_cfs, unlist = T)
}
\arguments{
\item{matches}{dataframe of match rows}

\item{field}{name of field of interest}

\item{schema}{function defining schema for column names}

\item{unlist}{(boolean) unlist the result? Vast majority of time TRUE, FALSE
if the content has complicated content not fit for a vector.}
}
\description{
Returns a dataframe with the results for every robot in matches for the field
specified in field.
}
\examples{
mil23 <- event_matches("2023mil")
get_robot_field_df(mil23, "autoChargeStation")
mar17 <- event_matches("2017mrcmp")
get_robot_field_df(mar17, "auto", schema = schema_csf)
}
