% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{get_team_stations}
\alias{get_team_stations}
\title{Get Team Stations}
\usage{
get_team_stations(matches, key)
}
\arguments{
\item{matches}{dataframe of matches with alliances broken out}

\item{key}{team key}
}
\value{
data frame with match indices and alliance station
}
\description{
Returns a df containing the indices of a team's matches and the alliance
station for that team in that match.
}
