% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{id2int}
\alias{id2int}
\title{ID to Int}
\usage{
id2int(id, prefix = "frc")
}
\arguments{
\item{id}{TBA-legal team formatted string (i.e. "frc3504")}

\item{prefix}{string prefix for id (in overwhelming majority of cases "frc")}
}
\description{
Coerce a string in team formatting to an int
}
