% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{id_unplayed}
\alias{id_unplayed}
\title{ID Unplayed Matches}
\usage{
id_unplayed(matches)
}
\arguments{
\item{matches}{a dataframe of TBA match objects}
}
\description{
This function takes a dataframe of TBA match objects and identifies matches
that have not been played.
}
\details{
We identify a match as unplayed if both red and blue alliances have
a score of -1.
}
