% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scout.R
\name{import_scout_json}
\alias{import_scout_json}
\title{Import json files from Scout exports}
\usage{
import_scout_json(jsonfile)
}
\arguments{
\item{jsonfile}{}
}
\description{
Import json files from Scout exports
}
