% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{initialize_scoutR}
\alias{initialize_scoutR}
\title{Initialize scoutR}
\usage{
initialize_scoutR(auth_key)
}
\description{
Function to start scoutR by writing the auth key for scoutR as well as
creating the documentation files for scoutR
}
