% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{is_event_key_official}
\alias{is_event_key_official}
\title{Identify an event key as official}
\usage{
is_event_key_official(key)
}
\arguments{
\item{key}{TBA event key}
}
\value{
(bool) Is the event official?
}
\description{
Returns a boolean describing whether a given event key is official
}
\examples{
is_event_key_official(read_event("2016parr"))
is_event_key_official(read_event("2016paphi"))
}
\author{
Gabriel Krotkov
}
