% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{is_event_official}
\alias{is_event_official}
\title{Identify an official event}
\usage{
is_event_official(event)
}
\arguments{
\item{event}{TBA Event object}
}
\value{
(bool) Is the input event official?
}
\description{
Returns a boolean describing whether an event is offical, using TBA's event
type code.
}
\examples{
is_event_official(read_event("2016parr"))
is_event_official(read_event("2016pahat"))
}
\author{
Gabriel Krotkov
}
