% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{is_qual_match}
\alias{is_qual_match}
\title{Identify qualification match}
\usage{
is_qual_match(match)
}
\arguments{
\item{match}{TBA Match object}
}
\value{
boolean
}
\description{
Returns a boolean given a match object representing whether the match was
a qualification match.
}
\examples{
paphi <- read_event_matches("2016paphi")
is_qual_match(paphi[[1]])
is_qual_match(paphi[[56]])
}
\author{
Gabriel Krotkov
}
