% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{is_score_posted}
\alias{is_score_posted}
\title{Score Posted Identified}
\usage{
is_score_posted(match)
}
\arguments{
\item{match}{TBA match object}
}
\value{
boolean, TRUE if the score is posted, FALSE otherwise
}
\description{
Returns a boolean of whether the score is posted given a match object
}
\examples{
match <- read_event_matches("2016paphi")[[1]]
is_score_posted(match)
}
\author{
Gabriel Krotkov
}
