% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_sf.c}
\alias{is_sf.c}
\title{Test if object is of class sf or sfc}
\usage{
is_sf.c(x)
}
\arguments{
\item{x}{a R object}
}
\value{
logical, whether or not the object if of class sf or sfc
}
\description{
Test if object is of class sf or sfc
}
