% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{lineup_design_matrix}
\alias{lineup_design_matrix}
\title{Lineup Design Matrix}
\usage{
lineup_design_matrix(matches)
}
\arguments{
\item{matches}{Dataframe of matches like output by event_matches}
}
\description{
Computes the lineup design matrix (indicator variables one-hot encoding each
robot's presence in a match). When used to fit a linear regression through
the intercept with scores as the response, the resulting coefficients
are equal to OPR. We use the nomenclature "lineups" to reflect NBA "lineup
data" which is related to the development of OPR/Calculated Contribution.
}
\details{
Assumes match order is irrelevant. Casts the final output to a
data.frame because the `lm` function expects a data.frame. Returns blue
alliances as a block, and then red alliances. We call this function the
"design matrix" because it only includes the indicator variables, and not
any of the responses.
}
\examples{
matches <- event_matches("2023mil", match_type = "qual")
matches <- matches[order(matches$match_number), ]
design <- lineup_design_matrix(matches)
design$score <- c(matches$blue_score, matches$red_score)
fit <- lm(score ~ 0 + ., data = design)
summary(fit) # retrieves OPRs
}
