% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_locations.R
\name{link_locations}
\alias{link_locations}
\title{Link event locations}
\usage{
link_locations(data, start_loc, end_loc)
}
\arguments{
\item{data}{A data frame with event location data.}

\item{start_loc}{A character vector identifying the \code{x} and \code{y}
coordinates for the start of an event in \code{data}.}

\item{end_loc}{A character vector identifying the \code{x} and \code{y}
coordinates for the end of an event in \code{data}.}
}
\value{
A simple feature object with a new variable named \code{geometry}.
The variable \code{geometry} is a linestring object.
}
\description{
This function converts \code{data} from a tibble to a simple feature object
by adding a linestring geometry that links the starting and ending locations
of each event in \code{data}.
}
\examples{
# load dplyr for examples
library(dplyr)

# read event data given in package
file_path <- system.file("extdata", "events_england.json", package = "scoutr")
events <- read_events(file_path)

events \%>\%
  select(event_sec:end_y) \%>\%
  transform_locations(x = c("start_x", "end_x"), y = c("start_y", "end_y")) \%>\%
  link_locations(start_loc = c("start_x", "start_y"), end_loc = c("end_x", "end_y"))

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
