% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{load_sync}
\alias{load_sync}
\title{Load Sync}
\usage{
load_sync()
}
\value{
Returns list with items "team_list", "rankings", "oprs", "alliances",
"match_schedule".
}
\description{
Returns synced data saved by tba_sync
}
\examples{
load_sync()
}
\author{
Gabriel Krotkov
}
