% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{LOOCV}
\alias{LOOCV}
\title{Leave-One-Out Cross Validation}
\usage{
LOOCV(fit)
}
\arguments{
\item{fit}{a fitted output of `lm` or `glm` or similar}
}
\value{
numeric estimate of leave-one-out CV risk
}
\description{
Computes the leave-one-out cross validation error of a regression fit object
}
\details{
Uses a "shortcut" formula for LOOCV. For details, see formula 5.2
in ISLR by James, Witten, Hastie, and Tibshirani (v2), pg 202
}
