% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markOutliers.R
\name{markOutliers
Creates new outlier columns in Seurat metadata slot for variables of interest.}
\alias{markOutliers
Creates new outlier columns in Seurat metadata slot for variables of interest.}
\alias{markOutliers}
\title{markOutliers}
\usage{
markOutliers(x, vars = c(...), nmad, bound)
}
\arguments{
\item{x}{A seurat object}

\item{vars}{A vector column names in Seurat object metadata slot to be tested}

\item{nmad}{The number of median absolute deviations to consider a cell an outlier.}
}
\value{
A list of seurat objects (one per sample)
}
\description{
Wrapper function for \code{scuttle::isOutlier}. Calculates outliers based on
median-average-deviations (MAD). Can be parallelized across multiple samples
using \code{purrr::map}
}
\examples{
# Call the function on a single Seurat object
outlierObj <- markOutliers(
  seuratObj,
  vars = c("nCount_RNA", "nFeature_RNA", "percentMT", "log10GenesPerUMI"),
  nmad = 5,
  bound = "higher"
)

# Apply the function across a list of Seurat objects using purrr::map
OutlierList <- purrr::map(
  samples,
  ~ markOutliers(
    .x,
    vars = c("nCount_RNA", "nFeature_RNA", "percentMT", "log10GenesPerUMI"),
    nmad = 5,
    bound = "higher"
  )
)

}
