% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markOutliers.R
\name{markOutliers}
\alias{markOutliers}
\title{markOutliers}
\usage{
markOutliers(x, vars = c(...), nmad, bound)
}
\arguments{
\item{x}{A seurat object}

\item{vars}{A vector column names in seurat object metadata slot to be tested for outliers}

\item{nmad}{The number of median absolute deviations to consider a cell an outlier. 3-5 is recommended. 3 is more stringent.}

\item{bound}{One of "higher", "lower", or "both"}
}
\value{
A list of seurat objects (one per sample)
}
\description{
Wrapper function for \code{scuttle::isOutlier}. If multiple samples need to be used, utilize \code{purrr::map}
}
\examples{
# Call the function on a single sample
outlierObj <- markoutliers(seuratObj, vars = c("nCount_RNA", "nFeature_RNA", "percentMT", "log10GenesPerUMI"), nmad = 5, bound = "higher")
# Call the function across multiple Seurat objects
OutlierList <- map(samples, ~ markOutliers(.x, vars = c("nCount_RNA", "nFeature_RNA", "percentMT", "log10GenesPerUMI"), nmad = 5, bound = "higher"))
}
