% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{match_keys}
\alias{match_keys}
\title{Extract match keys}
\usage{
match_keys(matches)
}
\arguments{
\item{matches}{List of TBA match objects}
}
\value{
Vector of match keys
}
\description{
Returns a vector of match keys given a list of TBA match objects
}
\examples{
event_key <- "2016paphi"
match_keys(read_event_matches(event_key))
}
\author{
Gabriel Krotkov
}
