% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{match_sb}
\alias{match_sb}
\title{Match (Statbotics)}
\usage{
match_sb(key)
}
\arguments{
\item{key}{match key (e.g. "2023milf1m3")}
}
\description{
Returns a list representing statbotics info for a particular match
}
