% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{matches_sb}
\alias{matches_sb}
\title{Matches (Statbotics)}
\usage{
matches_sb(params)
}
\arguments{
\item{params}{Named list of parameters to pass to the statbotics API call.}
}
\description{
Matches (Statbotics)
}
