% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{max_coprs}
\alias{max_coprs}
\title{Max cOPRs}
\usage{
max_coprs(tms, yr, ...)
}
\arguments{
\item{tms}{(numeric) vector of team ids}

\item{yr}{(numeric) year}

\item{...}{additional arguments passed to team_max_coprs()}
}
\description{
Max cOPRs
}
