% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{max_oprs}
\alias{max_oprs}
\title{Max OPRs}
\usage{
max_oprs(tms, yr, ...)
}
\arguments{
\item{tms}{(numeric) vector of team ids}

\item{yr}{(numeric) year}

\item{...}{additional arguments passed to team_max_oprs()}
}
\description{
Max OPRs
}
