% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMetadata.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{mergeMetadata}
\usage{
mergeMetadata(x)
}
\arguments{
\item{x}{A named-list of Seurat objects}
}
\value{
A dataframe of merged metadata
}
\description{
Efficiently merge metadata across multiple Seurat objects into one master dataframe.
}
\examples{
# list of seurat objects with names
objects <- c("obj1", "obj2", "obj3", "obj4")
# Vector of sample IDs
sampleIDs <- c("S1", "S2", "S3", "S4")
names(objects) <- sampleIDs
metadata <- mergeMetadata(objects)

}
