% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{multievent_insights}
\alias{multievent_insights}
\title{Multievent Insights}
\usage{
multievent_insights(keys, insight_as_col = FALSE)
}
\arguments{
\item{keys}{vector or list of TBA event keys (ex. "2023txfor")}

\item{insight_as_col}{(bool)}
}
\description{
Creates a single insights dataframe for a vector of event keys
}
