% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{normalize_weights}
\alias{normalize_weights}
\title{Normalize Weights}
\usage{
normalize_weights(w, len_out = NA)
}
\arguments{
\item{w}{vector of weights}

\item{len_out}{positive integer, desired output length of the normed weights}
}
\value{
vector of scaled weights, guaranteed to be integer
}
\description{
Takes a vector of real, non-negative weights and converts them to an integer
form while maintaining the ratios between all elements.
}
\details{
Iteratively gets the LCM of all the denominators, so we are
guaranteed to get integer results after multiplying by that LCM. And, because
it is the *least* common multiple, these will be the smallest possible
integer weights.
}
