% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opr_swings}
\alias{opr_swings}
\title{OPR Swing data (2005 - 2025)}
\format{
An object of class \code{data.frame} with 95207 rows and 5 columns.
}
\source{
thebluealliance.com
}
\usage{
opr_swings
}
\description{
"mean_diff" represents the mean change in standardized OPR coefficients
(mean across all teams at the event) after the corresponding match
"ppt" (plays per team) is the number of match plays the event has had,
divided by the number of teams at the event. So for example after 2025mdsev
ran match 13, there had been (13 * 6 = ) 78 "plays", across 28 teams at the
event, for about 2.79 ppt.
}
\keyword{datasets}
