% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perturbs}
\alias{perturbs}
\title{Yearly OPR perturbation data (2005 - 2025)}
\format{
An object of class \code{data.frame} with 95207 rows and 5 columns.
}
\source{
thebluealliance.com
}
\usage{
perturbs
}
\description{
"mean_diff" represents the mean change in standardized OPR coefficients
(mean across all teams at the event) after the corresponding match
}
\keyword{datasets}
