% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{plotOutliers}
\usage{
plotOutliers(metadata, variable, outlierData, outDir, width, height)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{variable}{A valid column in the metadata dataframe}

\item{outlierData}{A column name indicating results of \code{scuttle::isOutlier} function}

\item{outDir}{A directory path ending in "/" to hold output files}

\item{width}{Plot width}

\item{height}{Plot height}
}
\value{
A ggplot object
}
\description{
Generates publication quality violin plots showing outlier
distributions for any QC metric. Can be used in conjunction
with \code{ggpubr::ggarrange} to plot individual outlier plots as well as a paneled plot.
}
\examples{
# Run after using `markOutliers`
plotOutliers(metadata = metadata, 
variable = "nCount_peaks", 
outlierData = "nCount_peaks_outliers", 
outDir = outputDir, 
width = 10, 
height = 10)

}
