% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_RNA_basic.R
\name{preprocess_RNA_basic}
\alias{preprocess_RNA_basic}
\title{preprocess_RNA_basic}
\usage{
preprocess_RNA_basic(x, ident, nCount_thresh = 300, mito_gene_prefix)
}
\arguments{
\item{x}{A seurat object}

\item{ident}{Ident corresponding to seurat object}

\item{nCount_thresh}{Threshold for filtering based on nCount_RNA (default = 300)}

\item{mito_gene_prefix}{\code{pattern} argument from \code{Seurat::PercentageFeatureSet} (i.e., "^mt-")}
}
\value{
xSub, a seurat object filtered for \code{nCount_thresh} cells with added metadata for percent mitochondria and log10 genes per UMI
}
\description{
Filters a seurat object based on nCount RNA and adds additional metadata for percent mitochondria and log10 genes per UMIs
}
\examples{
# Run on a single seurat object
subset <- preprocess_RNA_basic(x, "Sample_1", nCount_thresh = 250, mito_gene_prefix = "^mt-")
}
