% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{prescout}
\alias{prescout}
\title{Prescout}
\usage{
prescout(event_code, breakdown = TRUE, manual_teams = NULL)
}
\arguments{
\item{event_code}{TBA-legal event code (ex. "2024paca")}

\item{breakdown}{(bool) Compute OPR and EPA for all possible response
variables? If FALSE, will only return topline values.}

\item{manual_teams}{(int) additional teams to manually add to the team list}
}
\description{
Given an event code, return a df with prescouting data
}
\details{
Checks for match duplication, which will stop execution if TRUE.
}
\examples{
gpr24 <- prescout("2024paca")
newton25 <- prescout("2025newton", manual_teams = c(1712, 6672, 3504))
}
