% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ridge.R
\name{pridge_lambda_cv}
\alias{pridge_lambda_cv}
\title{Prior Ridge Lambda Cross Validation (foreach version)}
\usage{
pridge_lambda_cv(
  design,
  response,
  priors,
  grid,
  plot_mses = TRUE,
  n_cores = NULL
)
}
\arguments{
\item{design}{(matrix, or coercable to matrix) 1-hot encoded team lineups,
matches on the rows, teams on the columns}

\item{response}{(vector) response vector corresponding to `design`;
typically alliance scores}

\item{priors}{(vector) length equal to ncol(design) representing a best
guess at the coefficient without match data. Typically pre-event EPA.}

\item{grid}{(vector) all lambda values (regularization parameter) to consider}

\item{plot_mses}{(boolean) if TRUE, output a plot showing the CV results}

\item{n_cores}{(int) the number of cores to parallelize over; or NULL to use
the max minus 1.}
}
\description{
Runs leave-one-out cross validation across a grid of lambda and returns the
MSEs across the whole grid. This version uses the 'foreach' package for
cleaner parallel execution.
}
