% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ridge.R
\name{pridge_loocv}
\alias{pridge_loocv}
\title{Prior Ridge leave-one-out cross validation}
\usage{
pridge_loocv(X, y, lambda, beta_0, mse = TRUE)
}
\arguments{
\item{X}{(matrix) columns are 1-hot encoded teams, each row an alliance-match}

\item{y}{(vector) response - typically alliance scores}

\item{lambda}{regularization parameter}

\item{beta_0}{vector of priors to regularize towards}
}
\description{
Compute prior ridge leave-one-out cross validation error for a specific
data matrix, response vector, regularization, and priors vector
}
