% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ridge.R
\name{prior_ridge}
\alias{prior_ridge}
\title{Prior Ridge}
\usage{
prior_ridge(X, y, lambda, beta_0)
}
\arguments{
\item{X}{(matrix) columns are 1-hot encoded teams, each row an alliance-match}

\item{y}{(vector) response - typically alliance scores}

\item{lambda}{regularization parameter}

\item{beta_0}{vector of priors to regularize towards}
}
\description{
Given a data matrix, response vector, regularization parameter, and vector
of priors, compute a prior ridge model
}
