% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcRidges.R
\name{qcRidges}
\alias{qcRidges}
\title{qcRidges}
\usage{
qcRidges(metadata, variable, sampleColors, figDir, width, height)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{variable}{A valid column name present in your metadata dataframe}

\item{sampleColors}{A color vector for samples}

\item{figDir}{A directory path ending in "/" to hold output files}
}
\value{
A ggplot object
}
\description{
Generates publication quality quality control ridgeplots. Requires that you concatenate all
sample metadata across multiple objects into one dataframe. Can be used in conjunction with \code{ggpubr::ggarrange}
to plot panelled figure of all QC metrics.
}
\examples{
# Generate individual plots as well as a combined one
arrangedViolins <- ggarrange(qcRidges(metadata, "pct_reads_in_peaks", logTransform = FALSE, colors, figDir),
qcRidges(metadata, "atac_peak_region_fragments", colors, figDir, width = 8, height = 10),
qcRidges(metadata, "TSS.enrichment", colors, figDir, width = 8, height = 10),
qcRidges(metadata, "nucleosome_signal", colors, figDir, width = 8, height = 10),
qcRidges(metadata, "nFeature_peaks", colors, figDir, width = 8, height = 10)
ggsave(paste0(figDir, "ATAC_Violin_Summary.png"), arrangedViolins, width = 16, height = 10)
}
