% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcScatter.R
\name{qcScatter}
\alias{qcScatter}
\title{qcScatter}
\usage{
qcScatter(
  metadata,
  Xvar,
  Yvar,
  logTransformX = FALSE,
  logTransformY = FALSE,
  colors,
  outDir,
  width,
  height
)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{logTransformX}{Logical, whether or not to log transform the X-axis (default = FALSE)}

\item{logTransformY}{Logical, whether or not to log transform the Y-axis (default = FALSE)}

\item{colors}{A color vector for samples}

\item{outDir}{A directory path ending in "/" to hold output files}

\item{width}{Plot width}

\item{height}{Plot height}

\item{XVar}{A valid column name present in your metadata dataframe to plot on the X-axis}

\item{YVar}{A valid column name present in your metadata dataframe to plot on the Y-axis}
}
\value{
A ggplot object
}
\description{
Generates publication quality quality control scatter plot for 2 metadata variables. Requires that you concatenate all
sample metadata across multiple objects into one dataframe.
}
\examples{
# Generate scatter plot
compareDists(metadata, Xvar = "log_atac_peak_region_fragments", Yvar = "pct_reads_in_peaks", logTransformX = FALSE, 
logTransformY = FALSE, colors = colors, outDir = figDir, width = 12, height = 10)
}
