% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcViolin.R
\name{qcViolin}
\alias{qcViolin}
\title{qcViolin}
\usage{
qcViolin(metadata, variable, logTransform, sampleColors, figDir)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{variable}{A valid column name present in your metadata dataframe}

\item{logTransform}{Logical, whether or not to log transform the y-axis}

\item{sampleColors}{A color vector for samples}

\item{figDir}{A directory path ending in "/" to hold output files}
}
\value{
A ggplot object
}
\description{
Generates publication quality quality control violins. Requires that you concatenate all
sample metadata across multiple objects into one dataframe. Can be used in conjunction with \code{ggpubr::ggarrange}
to plot panelled figure of all QC metrics.
}
\examples{
# Generate individual plots as well as a combined one
arrangedViolins <- ggarrange(qcVioli(metadata, "pct_reads_in_peaks", logTransform = FALSE, colors, figDir),
qcViolin(metadata, "atac_peak_region_fragments", logTransform = TRUE, colors, figDir),
qcViolin(metadata, "TSS.enrichment", logTransform = FALSE, colors, figDir),
qcViolin(metadata, "nucleosome_signal", logTransform = FALSE, colors, figDir),
qcViolin(metadata, "nFeature_peaks", logTransform = FALSE, colors, figDir))
ggsave(paste0(figDir, "ATAC_Violin_Summary.png"), arrangedViolins, width = 16, height = 10)
}
