% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{qual_schedule}
\alias{qual_schedule}
\title{Event Schedule}
\usage{
qual_schedule(event_code, int_output = TRUE)
}
\arguments{
\item{event_code}{TBA-legal event key}

\item{int_output}{convert TBA keys (e.g. "frc1712") to int ("1712")?}
}
\description{
Returns a dataframe with the event's qualification match schedule
}
