% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_ATAC.R
\name{read_10x_ATAC}
\alias{read_10x_ATAC}
\title{read_10x_ATAC}
\usage{
read_10x_ATAC(
  sample,
  files_dir,
  annotations,
  min.cells,
  min.features,
  blacklist,
  figDir
)
}
\arguments{
\item{sample}{A sample name (must be same name as your cellranger outputs to dynamically construct file paths)}

\item{files_dir}{A directory path ending in "/" where raw 10x data lives}

\item{annotations}{A GRanges object with organism annotations}

\item{min.cells}{Minimum number of cells a feature needs to be present in to be retained}

\item{min.features}{Minimum number of features needed in a cell to be retained}

\item{blacklist}{A blacklist region file in bed format}

\item{figDir}{A directory path ending in "/" where density plots will be output to}
}
\value{
A list of seurat objects (one per sample)
}
\description{
When multiple samples are present, \code{read_10x_ATAC} can read each sample's
filtered feature barcode matrix h5 file and generate a \code{ChromatinAssay} and
\code{Seurat} object for each. Each of these objects is then combined into a list
of objects. This function can be easily parallelized using \code{purrr::map}.
This function requires that you have your organism annotations as a \code{GRanges}
object and blacklist region file in bed format. By default, it will output
density plots of nCount peaks vs TSS enrichment.
}
\examples{
# Path to sample directory and output directory
wd <- "/my/path/to/data/"
outDir <- "/my/path/to/data/outputs/"
# Vector of sample IDs
sampleIDs <- c("S1", "S2", "S3")
# Vector of idents
idents <- c("A", "B", "C")
# Read in annotations and blacklist bed
anno <- readRDS("path/to/annotations.Rds")
blacklist <- read.csv("/path/to/blacklist.bed", sep = "\t")
# Call the function
seurat_list <- map(samples, 
  ~ run_atac_qc_filtered(
    .x, 
    wd, 
    anno, 
    min.cells = 10, 
    min.features = 200, 
    blacklist, 
    outDir)
    )

}
