% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_GEX.R
\name{read_10x_GEX}
\alias{read_10x_GEX}
\title{read_10x_GEX}
\usage{
read_10x_GEX(sampleDir, sample, ident, min.cells, min.features)
}
\arguments{
\item{sampleDir}{Path to the sample directory containing multiple subfolders (one per sample)}

\item{sample}{A vector of sample names that correspond to how subfolders in \code{sampleDir} are named}

\item{ident}{A vector of desired ident names for seurat object. Idents must correspond to same order in \code{sample}}

\item{min.cells}{Minimum number of cells a feature needs to be present in to be retained}

\item{min.features}{Minimum number of features needed in a cell to be retained}
}
\value{
A list of seurat objects (one per sample)
}
\description{
When multiple samples are present, \code{read_10x_GEX} can read each sample's filtered feature barcode matrix h5 file
and generate a seurat object for each. Each of these seurat objects is then combined into a list of objects.
This function can be easily parallelized using \code{purrr::map}
}
\examples{
# Path to sample directory
wd <- "/my/path/to/data/"
# Vector of sample IDs
sampleIDs <- c("S1", "S2", "S3")
# Vector of idents
idents <- c("A", "B", "C")
# Call the function
seuratList <- map(samples, ~ read_10x_multiGEX(.x, ident = idents, min.cells = 10, min.features = 100))
}
