% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_multi.R
\name{read_10x_multi}
\alias{read_10x_multi}
\title{read_10x_multi}
\usage{
read_10x_multi(sampleDir, sample, ident)
}
\arguments{
\item{sampleDir}{Path to the sample directory containing multiple subfolders (one per sample)}

\item{sample}{A vector of sample names that correspond to how subfolders in \code{sampleDir} are named}

\item{ident}{A vector of desired ident names for seurat object. Idents must correspond to same order in \code{sample}}
}
\value{
A list of seurat objects (one per sample)
}
\description{
When multiple samples are present, \code{read_10x_multi} can read each sample's filtered feature barcode matrix h5 file
and generate a seurat object for each. Each of these seurat objects is then combined into a list of objects.
This function can be easily parallelized using \code{purrr::map}
}
\examples{
# Path to sample directory
wd <- "/my/path/to/data/"
# Vector of sample IDs
sampleIDs <- c("S1", "S2", "S3")
# Vector of idents
idents <- c("A", "B", "C")
# Call the function
seuratList <- purrr::map(sampleIDs, read_10x_multi(wd, sampleIDs, idents))
}
