% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_district_events}
\alias{read_district_events}
\title{Read District Events}
\usage{
read_district_events(district_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{district_key}{(character) TBA identifier of district in a year}

\item{simple}{(bool) simplify event objects?}
}
\value{
list of district event objects
}
\description{
Returns a list of district events given district key
}
\examples{
read_district_events("2022fit")
}
\author{
Gabriel Krotkov
}
