% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_district_rankings}
\alias{read_district_rankings}
\title{Read District Rankings}
\usage{
read_district_rankings(district_key)
}
\arguments{
\item{district_key}{(character) district identifier}
}
\value{
list of district rankings objects
}
\description{
Gets the rankings of an input district
}
\examples{
read_district_rankings("2016mar")
}
\author{
Gabriel Krotkov
}
