% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_district_teams}
\alias{read_district_teams}
\title{Read District Teams}
\usage{
read_district_teams(district_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{district_key}{(character) TBA key for district}

\item{simple}{(bool) simplify team objects?}
}
\value{
list of district teams
}
\description{
Get a list of all teams in a district
}
\examples{
read_district_teams("2022fit")
read_district_teams("2016mar")
}
\author{
Gabriel Krotkov
}
