% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_districts}
\alias{read_districts}
\title{Read Districts}
\usage{
read_districts(year = YEAR)
}
\arguments{
\item{year}{(numeric/character)}
}
\value{
List of district objects
}
\description{
Reads the districts in FIRST for the given year
}
\examples{
read_districts(2022)
}
\author{
Gabriel Krotkov
}
