% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event}
\alias{read_event}
\title{Read an event object from a key}
\usage{
read_event(key, simple = FALSE)
}
\arguments{
\item{key}{TBA legal event key}

\item{simple}{(bool) return simple version?}
}
\value{
TBA Event Object
}
\description{
Returns an event object given the key
}
\examples{
read_event("2016paphi")
read_event("2016mrcmp", simple = TRUE)
}
\author{
Gabriel Krotkov
}
