% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_alliances}
\alias{read_event_alliances}
\title{Read event alliances from key}
\usage{
read_event_alliances(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA Alliances object
}
\description{
Returns an event alliances object given an event key
}
\examples{
read_event_alliances("2015paphi")
}
\author{
Gabriel Krotkov
}
