% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_awards}
\alias{read_event_awards}
\title{Read Event Awards}
\usage{
read_event_awards(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA Awards object
}
\description{
Returns event awards given event key
}
\examples{
read_event_awards("2017pawch")
}
\author{
Gabriel Krotkov
}
