% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_coprs}
\alias{read_event_coprs}
\title{Read Event cOPRs}
\usage{
read_event_coprs(key)
}
\arguments{
\item{key}{(string) event key}
}
\description{
Reads event cOPRs from the event_insights tab and returns them as a list.
}
\author{
Dr. Holt Oliver
}
