% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_district_points}
\alias{read_event_district_points}
\title{Read Event District Points}
\usage{
read_event_district_points(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA district points event object
}
\description{
Returns event district points given event key
}
\examples{
read_event_district_points("2016mrcmp")
}
\author{
Gabriel Krotkov
}
