% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_insights}
\alias{read_event_insights}
\title{Read Event Insights from key}
\usage{
read_event_insights(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA Insights object
}
\description{
Returns event insights given event key
}
\examples{
read_event_insights("2018pawch")
}
\author{
Gabriel Krotkov
}
