% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_matches}
\alias{read_event_matches}
\title{Read Event Matches}
\usage{
read_event_matches(event_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{event_key}{(string) TBA legal event key}

\item{simple}{(bool) simple format response?}

\item{keys}{(bool) return only keys?}
}
\value{
Returns list of matches
}
\description{
Given an event key returns a list of all matches at the event
}
\examples{
read_event_matches("2016mrcmp")
read_event_matches("2014mrcmp", keys = TRUE)
}
\author{
Gabriel Krotkov
}
