% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_oprs}
\alias{read_event_oprs}
\title{Read Event OPRs}
\usage{
read_event_oprs(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA OPRs object
}
\description{
Returns event oprs given event key
}
\examples{
read_event_oprs("2018mrcmp")
}
\author{
Gabriel Krotkov
}
