% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_playoff_matches}
\alias{read_event_playoff_matches}
\title{Read Event Playoff Matches}
\usage{
read_event_playoff_matches(event_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{event_key}{TBA legal event key}

\item{simple}{return simple format match objects?}

\item{key}{return keys only?}
}
\value{
List of TBA Match objects, or vector of keys if keys = TRUE
}
\description{
Reads an event's playoff matches given the event key
}
\examples{
read_event_playoff_matches("2018pawch")
}
\author{
Gabriel Krotkov
}
