% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_predictions}
\alias{read_event_predictions}
\title{Read Event Predictions}
\usage{
read_event_predictions(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA Predictions object
}
\description{
Returns event predictions given event key
}
\examples{
read_event_predictions("2018mrcmp")
}
\author{
Gabriel Krotkov
}
