% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_qual_matches}
\alias{read_event_qual_matches}
\title{Read Event Qualification Matches}
\usage{
read_event_qual_matches(event_key, simple = FALSE, keys = FALSE)
}
\arguments{
\item{event_key}{TBA legal event key}

\item{simple}{(bool) simple format response?}

\item{keys}{(bool) restrict response to only keys?}
}
\value{
List of qualification matches
}
\description{
Given an event key, returns the event's qualification matches
}
\examples{
read_event_qual_matches("2016hop", simple = TRUE)
}
\author{
Gabriel Krotkov
}
