% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_rankings}
\alias{read_event_rankings}
\title{Read Event Rankings}
\usage{
read_event_rankings(key)
}
\arguments{
\item{key}{TBA legal event key}
}
\value{
TBA rankings object
}
\description{
Returns event rankings given event key
}
\examples{
read_event_rankings("2015paphi")
}
\author{
Gabriel Krotkov
}
