% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_event_teams}
\alias{read_event_teams}
\title{Read Event Teams}
\usage{
read_event_teams(event_key, statuses = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{statuses}{(boolean) get team statuses?}

\item{simple}{(boolean) return simple format?}

\item{keys}{(boolean) return keys only?}

\item{key}{TBA legal event key}
}
\value{
List of team objects from the given event
}
\description{
Returns event teams given event key. Either key or simple overrides statuses
}
\examples{
read_event_teams("2016hop")
read_event_teams("2016hop", statuses = TRUE)
read_event_teams("2016hop", simple = TRUE)
read_event_teams("2016hop", keys = TRUE)
}
\author{
Gabriel Krotkov
}
