% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events.R
\name{read_events}
\alias{read_events}
\title{Read Wyscout soccer events data}
\usage{
read_events(file, tidy_tibble = TRUE, keep_tags = TRUE)
}
\arguments{
\item{file}{File path to \code{events} JSON data.}

\item{tidy_tibble}{Logical value indicating if the returned object should be
a tidy tibble or not. If \code{FALSE}, a list is returned.}

\item{keep_tags}{Logical value indicating if \code{tag} variables should be
included in the returned tidy tibble or not. If \code{tidy_tibble} is
\code{FALSE}, \code{tag} variables are always included.}
}
\value{
A tidy tibble or list depending on the value of argument
\code{tidy_tibble}. A detailed description of the variables is given
at the reference cited below.
}
\description{
This function reads and processes the Wyscout \code{events} data.
}
\note{
These files are typically the largest and take the longest to process.
Expect it to take around two minutes to read in a year of event data for
a single league. The variable names in the returned tidy tibble are in
snake case, but they are analogous to those used by Wyscout.

The JSON data, \code{events_england.json}, included in this package is only
a single game of events subsetted from the full publicly available
Wyscout England event data.
}
\examples{
file_path <- system.file("extdata", "events_england.json", package = "scoutr")

events <- read_events(file_path)
events_list <- read_events(file_path, tidy_tibble = FALSE)

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
\seealso{


Other read functions: 
\code{\link{read_coaches}()},
\code{\link{read_competitions}()},
\code{\link{read_matches}()},
\code{\link{read_players}()},
\code{\link{read_teams}()}
}
\concept{read functions}
