% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_match_statbotics}
\alias{read_match_statbotics}
\title{Read matches}
\usage{
read_match_statbotics(key)
}
\arguments{
\item{key}{match key}
}
\description{
Returns match objects given a match key
}
