% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_match_zebra}
\alias{read_match_zebra}
\title{Zebra reader
Reads Zebra match data}
\usage{
read_match_zebra(match_key)
}
\arguments{
\item{match_key}{TBA match key}
}
\value{
Zebra match for the input match key, or NULL
}
\description{
Zebra reader
Reads Zebra match data
}
\examples{
read_match_zebra("2022txcmp2_qf4m3")
}
\author{
Dr. Holt Oliver
}
