% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_official_events}
\alias{read_official_events}
\title{Read Official Events}
\usage{
read_official_events(
  year = YEAR,
  trim_parents = TRUE,
  simple = FALSE,
  keys = FALSE
)
}
\arguments{
\item{year}{year of interest}

\item{trim_parents}{remove events with divisions below them?}

\item{simple}{return simple event objects?}

\item{keys}{return vector of keys?}
}
\value{
TBA event objects list, or vector if keys = TRUE
}
\description{
Reads official events from the TBA API.
}
\examples{
read_official_events(year = 2016, trim_parents = TRUE, keys = TRUE)
}
\author{
Gabriel Krotkov
}
