% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_team}
\alias{read_team}
\title{Read Team}
\usage{
read_team(n, simple = FALSE)
}
\arguments{
\item{n}{team number}

\item{simple}{- (boolean) FALSE if you want all available information}
}
\value{
Team object
}
\description{
Reads team object from numeric input
}
\examples{
read_team(1712)
read_team(708)
read_team(6672)
}
\author{
Gabriel Krotkov
}
