% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_team_awards}
\alias{read_team_awards}
\title{Read Team Awards}
\usage{
read_team_awards(team, year = NA, event = NA)
}
\arguments{
\item{team}{(numeric/character) team of interest}

\item{year}{(numeric/character) year of interest}

\item{event}{(character) TBA legal event key}
}
\value{
List of award objects
}
\description{
Reads the awards that the given team has won
}
\examples{
read_team_awards(1712, year = 2018)
}
\author{
Gabriel Krotkov
}
