% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_team_events}
\alias{read_team_events}
\title{Read Team Events}
\usage{
read_team_events(n, year = NA, official = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{n}{(numeric) team number}

\item{year}{(numeric/character) year of interest}
}
\value{
list of event objects
}
\description{
Returns a team's event given the team number
}
\examples{
read_team_events(1712, year = 2018, simple = TRUE)
read_team_events(1712, keys = TRUE)
}
\author{
Gabriel Krotkov
}
