% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_team_matches}
\alias{read_team_matches}
\title{Read Team Matches}
\usage{
read_team_matches(
  team,
  year = NA,
  event = NA,
  official = FALSE,
  simple = FALSE,
  keys = FALSE
)
}
\arguments{
\item{team}{(numeric/character) team key or number}

\item{year}{(numeric) year of interest, or NA if all}

\item{event}{(character) TBA event key}

\item{official}{(bool) restrict to official events?}

\item{simple}{(bool) present simple data objects?}

\item{keys}{(bool) return keys only?}
}
\value{
list of match objects with the given team
}
\description{
Returns all the team's matches
}
\examples{
read_team_matches(1712, year = 2016, keys = TRUE)
}
\author{
Gabriel Krotkov
}
