% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_team_matches_statbotics}
\alias{read_team_matches_statbotics}
\title{Read Team Matches}
\usage{
read_team_matches_statbotics(
  team = NA,
  year = NA,
  event = NA,
  week = NA,
  match = NA,
  elims = NA,
  metric = NA,
  ascending = NA,
  limit = NA,
  offset = NA
)
}
\arguments{
\item{team}{team number}

\item{year}{year of interest}

\item{event}{event code (ex. 2018pahat)}

\item{week}{(int) week of competition}

\item{match}{match code}

\item{metric}{statbotics metric to sort the resulting list by}

\item{ascending}{(bool)}

\item{limit}{(int) maximum length of the desired output.}

\item{offset}{(int) Length of the offset of the beginning of the pull from
the sorted list.}
}
\description{
Returns team-match objects based on supplied parameters.
}
