% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_team_official_matches}
\alias{read_team_official_matches}
\title{Read Official Team Matches}
\usage{
read_team_official_matches(team, year, event, simple, keys)
}
\arguments{
\item{team}{(numeric/character)}

\item{year}{(numeric) year of interest}

\item{event}{(character) event key}

\item{simple}{(bool) return simple objects?}

\item{keys}{(bool) return keys only?}
}
\value{
List of official matches by the team of interest
}
\description{
Returns a team's official matches
}
\examples{
read_team_official_matches(1712, 2015, NA, TRUE, TRUE)
}
\author{
Gabriel Krotkov
}
