% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_team_statbotics}
\alias{read_team_statbotics}
\title{Read Team}
\usage{
read_team_statbotics(team)
}
\arguments{
\item{team}{the team number}
}
\value{
List of team's stats on statbotics
}
\description{
Gets a statbotics team record
}
\examples{
read_team(6672)
read_team(1712)
read_team(708)
}
