% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_team_year_statbotics}
\alias{read_team_year_statbotics}
\title{Read Team Year}
\usage{
read_team_year_statbotics(n, year)
}
\arguments{
\item{n}{(int or character) team number}

\item{year}{(int or character) year}
}
\value{
list object of team-year data
}
\description{
Returns a "team-year" object representing the data of a team in a given year.
}
\examples{
read_team_year(1712, 2016)
}
