% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_team_years_statbotics}
\alias{read_team_years_statbotics}
\title{Read Team Years}
\usage{
read_team_years_statbotics(
  team = NA,
  year = NA,
  country = NA,
  district = NA,
  state = NA,
  metric = NA,
  ascending = NA,
  limit = NA,
  offset = NA
)
}
\arguments{
\item{team}{team number}

\item{year}{year of interest}

\item{country}{full name of country (ex. Canada)}

\item{district}{lowercase district abbreviation (ex. fma)}

\item{state}{uppercase abbreviation for state/province (ex. CA, TX)}

\item{metric}{statbotics metric to sort the resulting list by}

\item{ascending}{(bool)}

\item{limit}{(int) maximum length of the desired output.}

\item{offset}{(int) Length of the offset of the beginning of the pull from
the sorted list.}
}
\description{
Returns team-year objects based on supplied parameters.
}
