% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{read_teams_statbotics}
\alias{read_teams_statbotics}
\title{Read Teams}
\usage{
read_teams_statbotics(
  metric = "team",
  country = NA,
  district = NA,
  state = NA,
  active = NA,
  ascending = TRUE,
  limit = 100,
  offset = 0
)
}
\arguments{
\item{metric}{(string) statbotics metric to sort by. Valid metrics include
country, district, state, active.}

\item{country}{Full name of the country of interest (ex. Canada, Israel, USA)}

\item{district}{lowercase abbreviation for the district of interest (ex. fma)}

\item{state}{Uppercase abbreviation for the state/territory of interest}

\item{active}{(bool) Restrict to active teams?}

\item{ascending}{(bool) Sort ascending? (False for sort descending)}

\item{limit}{(int) maximum length of the desired output.}

\item{offset}{(int) Length of the offset of the beginning of the pull from
the sorted list.}
}
\value{
List of team objects with statbotics information
}
\description{
Reads a group of teams with a host of optional parameters. If a parameter is
left as NA, it is removed from the request to statbotics.
}
