% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_year_events}
\alias{read_year_events}
\title{Read Year Events}
\usage{
read_year_events(year = YEAR, simple = FALSE, keys = FALSE)
}
\arguments{
\item{year}{(int) year of interest}

\item{simple}{(bool) simple format return value?}

\item{keys}{(bool) return keys only?}
}
\value{
List of event objects in the given year
}
\description{
Returns all events for a given year. Keys overrides simple.
}
\examples{
read_year_events(year = 2014)
read_year_events(year = 2015, simple = TRUE)
read_year_events(year = 2016, keys = TRUE)
read_year_events(year = 2017, simple = TRUE, keys = TRUE)
}
\author{
Gabriel Krotkov
}
