% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{remove_unscored_matches}
\alias{remove_unscored_matches}
\title{Remove Unscored Matches}
\usage{
remove_unscored_matches(matches)
}
\arguments{
\item{matches}{List of TBA match objects}
}
\value{
List of TBA match objects without unposted matches
}
\description{
Returns a list of match objects without any matches without the score posted.
}
\examples{
matches <- read_event_matches("2022azfl")
remove_unscored_matches(matches)
}
\author{
Gabriel Krotkov
}
