% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_each_len}
\alias{rep_each_len}
\title{Repeat Each for Given Length}
\usage{
rep_each_len(v, len_out)
}
\arguments{
\item{v}{input vector to extend}

\item{len_out}{the desired final length of the vector}
}
\description{
Extend a vector to have a given (`len_out`) length by repeating elements
using the `each` functionality of `rep()`
}
