% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{schema_cfs}
\alias{schema_cfs}
\title{Schema for (color)_(field)Robot(station_num)}
\usage{
schema_cfs(color, station_num, field)
}
\arguments{
\item{color}{alliance color, either "red" or "blue"}

\item{station_num}{number of the driver station}

\item{field}{name of the relevant field}
}
\description{
This schema was the default between 2018 and 2023 (at least.)
cfs references color/field/station, the order of the terms
}
