% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{schema_csf}
\alias{schema_csf}
\title{Schema for (color)_robot(station_num)(Field)}
\usage{
schema_csf(color, station_num, field)
}
\arguments{
\item{color}{alliance color, either "red" or "blue"}

\item{station_num}{number of the driver station}

\item{field}{name of the relevant field}
}
\description{
This schema was the default in 2016 and 2017.
csf references color/station/field, the order of the terms
}
