% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_import.R
\name{scout_import_json}
\alias{scout_import_json}
\title{Import json files from Scout exports}
\usage{
scout_import_json(jsonfile)
}
\arguments{
\item{jsonfile}{the path to the json file to be imported}
}
\value{
a {sf} object using WGS 84 (CRS 4326) geographic coordinate system
}
\description{
Import json files containing geographic data and convert them to {sf} objects
}
\details{
Geographic data (waypoints and records) are saved in json files in SCOUT
exports. This function use the {jsonlite} package to import them as data
frames and then convert them in simple features objects using the {sf}
package.
}
