% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{simkeys}
\alias{simkeys}
\title{Simkeys}
\usage{
simkeys(req, simple = FALSE, keys = FALSE)
}
\arguments{
\item{req}{Request string}

\item{simple}{Boolean - attach simple?}

\item{keys}{Boolean - attach keys?}
}
\value{
Modified request string
}
\description{
Attaches simple or keys to the end of a request string
}
\details{
Note that keys = TRUE overrides simple = TRUE
}
\examples{
simkeys("", TRUE, FALSE)
simkeys("", FALSE, TRUE)
simkeys("", TRUE, TRUE)
}
\author{
Gabriel Krotkov
}
