% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{substr_right}
\alias{substr_right}
\title{Substring Right}
\usage{
substr_right(s, n)
}
\arguments{
\item{s}{input string}

\item{n}{number of characters from the right to include (1 indexed)}
}
\description{
Returns the n rightmost characters in s.
}
