% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substr_right_inv}
\alias{substr_right_inv}
\title{Inverted Substring Right}
\usage{
substr_right_inv(s, n)
}
\arguments{
\item{s}{input string}

\item{n}{number of characters (1 indexed) to chop off the right.}
}
\description{
Returns string s without the n rightmost characters
}
