% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ta_hexafecta24}
\alias{ta_hexafecta24}
\title{Team Attributes Hexafecta}
\format{
An object of class \code{data.frame} with 8420 rows and 2 columns.
}
\source{
thebluealliance.com
}
\usage{
ta_hexafecta24
}
\description{
DF noting whether each team has achieved the Team Attributes Hexafecta
as of 2024.
}
\keyword{datasets}
