% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{tba_sync}
\alias{tba_sync}
\title{TBA Sync}
\usage{
tba_sync(event_key)
}
\arguments{
\item{event_key}{TBA legal event key}
}
\value{
No return value, directly writes to tbasync.rda
}
\description{
Download critical TBA data for an event and saves it locally, clearing
previous cache. Intended to be run periodically by stepping outside the venue
at a competition.
}
\examples{
tba_sync("2022txirv")
tba_sync("2017pawch")
}
\author{
Gabriel Krotkov
}
