% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{team}
\alias{team}
\title{Team}
\usage{
team(key, simple = FALSE)
}
\arguments{
\item{key}{TBA legal team key}

\item{simple}{(bool) simplify object?}
}
\value{
Team object in list format
}
\description{
Read a team object from the TBA API
}
\author{
Gabriel Krotkov
}
