% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{team_awards}
\alias{team_awards}
\title{Team Awards}
\usage{
team_awards(key, year = NULL, event = NULL, recipients = FALSE)
}
\arguments{
\item{key}{TBA legal team key}

\item{year}{year of interest}

\item{event}{event of interest}

\item{recipients}{(bool) break out recipients column?}
}
\value{
tidy tibble of award history
}
\description{
Gets team award history
}
\examples{
team_awards("1712", year = 2016)
team_awards("frc1712", year = 2010, recipients = TRUE)
team_awards(1712, event = "2017pawch")
}
\author{
Gabriel Krotkov
}
