% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_event_sb}
\alias{team_event_sb}
\title{Team Event (Statbotics)}
\usage{
team_event_sb(tm, event_key)
}
\arguments{
\item{tm}{team number}

\item{event_key}{unique event key (e.g. "2025mdpas")}
}
\description{
Returns a 1-row dataframe
}
