% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{team_events}
\alias{team_events}
\title{Team Events}
\usage{
team_events(key, year = NA, official = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{key}{TBA legal team key. Can be in int or string format.}

\item{year}{year of interest}

\item{official}{(bool) get only official events?}

\item{simple}{(bool) simplify event objects?}

\item{keys}{(bool) return keys only?}
}
\value{
tidy tibble of event objects, or vector if keys = TRUE
}
\description{
Read events for a certain team
}
\examples{
team_events(1712, year = 2017)
team_events("1712", year = 2018, official = TRUE)
team_events("frc1712", keys = TRUE)
}
\author{
Gabriel Krotkov
}
