% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_events_sb}
\alias{team_events_sb}
\title{Team-Events (Statbotics)}
\usage{
team_events_sb(...)
}
\arguments{
\item{...}{Parameters passed to the statbotics /v3/team_events endpoint}
}
\description{
Returns a batch of statbotics team-event objects based on the given parameters.
}
\examples{
team_events_sb(team = 4821, year = 2025)
team_events_sb(event = "2025mdsev")
}
