% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{team_matches}
\alias{team_matches}
\title{Team Matches}
\usage{
team_matches(
  key,
  year = NA,
  event = NA,
  official = FALSE,
  simple = FALSE,
  keys = FALSE,
  alliances = TRUE,
  breakdown = TRUE,
  unplayed = FALSE
)
}
\arguments{
\item{key}{TBA legal team key}

\item{year}{year of interest}

\item{event}{TBA event key of interest}

\item{official}{(bool) get only official matches?}

\item{simple}{(bool) simplify match objects?}

\item{keys}{(bool) get only keys?}

\item{alliances}{(bool) break out alliance column in match objects?}

\item{breakdown}{(bool) break out score breakdown column?}

\item{unplayed}{(boolean) include matches with scores of -1 (indicating that the match has not been played?)}
}
\value{
tidy tibble of team matches, or vector if keys = TRUE
}
\description{
Team Matches
}
\examples{
team_matches(1712, keys = TRUE)
team_matches(1712, year = 2016)
team_matches(1712, event = "2015paphi")
team_matches(1712, official = TRUE, alliances = FALSE, breakdown = FALSE)
}
\author{
Gabriel Krotkov
}
