% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_matches_sb}
\alias{team_matches_sb}
\title{Team-Matches (Statbotics)}
\usage{
team_matches_sb(...)
}
\arguments{
\item{...}{Parameters passed to the statbotics /v3/team_matches endpoint}
}
\description{
Returns a batch of statbotics team-match objects based on the given parameters.
}
\examples{
team_matches_sb(team = 449, year = 2025)
}
