% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{team_participation_years}
\alias{team_participation_years}
\title{Team Participation Years}
\usage{
team_participation_years(n)
}
\arguments{
\item{n}{team number (either as number or string)}
}
\value{
vector of years
}
\description{
Returns a list of all years where the given team participated in at least
one event. Because this function is so simple, I won't clutter the namespace
by making a tba_readR and tba_tidyR version.
}
\examples{
team_participation_years(1712)
team_participation_years("3504")

}
