% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_sb}
\alias{team_sb}
\title{Team (Statbotics)}
\usage{
team_sb(tm)
}
\arguments{
\item{tm}{(int/chr) team number}
}
\description{
Returns a 1-row dataframe representing a team's statbotics results
}
\examples{
dawgma <- team_sb(1712)
blair <- team_sb(449)
}
