% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_year_sb}
\alias{team_year_sb}
\title{Team Year (Statbotics)}
\usage{
team_year_sb(tm, yr)
}
\arguments{
\item{tm}{(int/chr) team number}

\item{yr}{(int/chr) year of interest}
}
\description{
Returns a 1-row dataframe representing a team's performance in a single year
based on Statbotics EPA
}
\examples{
team_year_sb(6672, 2023)
team_year_sb(3504, 2024)
}
