% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{team_years_sb}
\alias{team_years_sb}
\title{Team-Years (Statbotics)}
\usage{
team_years_sb(...)
}
\arguments{
\item{...}{Parameters passed to the statbotics /v3/team_years endpoint}
}
\description{
Returns a batch of statbotics team-year objects based on the given parameters.
}
\examples{
team_years_sb(year = 2025, district = "chs")
team_years_sb(team = 449)
}
