% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{teams}
\alias{teams}
\title{Read Teams}
\usage{
teams(page_num, year = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{page_num}{index of desired page}

\item{year}{year of interest. If FALSE, gets all years}

\item{simple}{(bool) simplify return objects?}

\item{keys}{(bool) return keys only?}
}
\value{
tidy tibble of teams
}
\description{
Get team list paginated by team # in 500s.
}
\examples{
teams(0)
teams(4)
teams(1, year = 2016)
teams(0, simple = TRUE)
teams(3, keys = TRUE)
}
