% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{teams_sb}
\alias{teams_sb}
\title{Teams (Statbotics)}
\usage{
teams_sb(..., yr = NULL, event = NULL, match = NULL)
}
\arguments{
\item{...}{Optional parameters passed to v3/{teams} endpoint}

\item{yr}{(optional) year of interest}

\item{event}{(optional) event of interest}

\item{match}{(optional) match of interest}
}
\description{
Returns a batch of team (or team-year, or team-event, or team-match) objects
conforming to the provided parameters.
}
