% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams_yearly}
\alias{teams_yearly}
\title{Teams by Year}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55843 rows and 8 columns.
}
\usage{
teams_yearly
}
\description{
Indicates each team-year competed (as of 2024)
}
\keyword{datasets}
