% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{tf}
\alias{tf}
\title{Team Formatting}
\usage{
tf(n)
}
\arguments{
\item{n}{Team identifier as numeric or string}
}
\value{
TBA-legal identifier
}
\description{
Coerces team identifiers into TBA-legal format
}
\examples{
tf(1712)
tf("frc1712")
tf("1712")
}
\author{
Gabriel Krotkov
}
