% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_alliances}
\alias{tidy_alliances}
\title{Tidy Alliances}
\usage{
tidy_alliances(raw, unpack_picks = FALSE)
}
\arguments{
\item{raw}{list of alliances objects as from read_event_alliances}

\item{unpack_picks}{(bool) split up picks list into columns?}
}
\value{
tidy tibble of alliances
}
\description{
Returns a tidy tibble of the alliances at an event given alliances object
}
\examples{
tidy_alliances(read_event_alliances("2019mrcmp"), unpack_picks = TRUE)
}
\author{
Gabriel Krotkov
}
