% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_coprs}
\alias{tidy_coprs}
\title{Tidy Event cOPRs}
\usage{
tidy_coprs(data)
}
\arguments{
\item{data}{list of cOPR data}
}
\description{
Turns list of event cOPRs (from read_event_coprs) to a tidy dataframe
}
\author{
Dr. Holt Oliver
}
