% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_district_rankings}
\alias{tidy_district_rankings}
\title{Tidy District Ranking}
\usage{
tidy_district_rankings(raw, event_detail = c("none", "separate", "breakdown"))
}
\arguments{
\item{raw}{JSON list output from tba_readR}

\item{event_detail}{(chr) how much detail do we want to break the results down into?}
}
\description{
Tidy District Ranking
}
\details{
If "none", will include the event results list in a column in the output
If "separate" will separate the events but not break down the points
If "breakdown" will break down points earning at each event to the most
granular level.
}
