% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_event}
\alias{tidy_event}
\title{Tidy Event}
\usage{
tidy_event(raw)
}
\arguments{
\item{raw}{TBA API event object}
}
\value{
tidy tibble row of an event object
}
\description{
Makes a single event object into a tidy tibble. Designed to be easily
rbindable, since a tibble for a single event in tidy format is a rare use
case.
}
\examples{
tidy_event(read_event("2017pawch"))
}
\author{
Gabriel Krotkov
}
