% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_events}
\alias{tidy_events}
\title{Tidy Events}
\usage{
tidy_events(raw)
}
\arguments{
\item{raw}{list of TBA event objects}
}
\value{
Tidy tibble of events
}
\description{
Returns a tidy tibble of event objects given a list of TBA event objects
}
\examples{
tidy_events(read_district_events("2016mar"))
}
\author{
Gabriel Krotkov
}
