% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_insights}
\alias{tidy_insights}
\title{Tidy Insights}
\usage{
tidy_insights(raw, insight_as_col = FALSE)
}
\arguments{
\item{raw}{Sublist of insight data, usually either quals or playoffs}

\item{insight_as_col}{(bool) Do we want to transpose the data? If TRUE,}
}
\description{
Function to handle JSON input of insight data, returning a tidy dataframe.
}
