% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_matches}
\alias{tidy_matches}
\title{Tidy Matches}
\usage{
tidy_matches(
  raw,
  alliances = FALSE,
  breakdown = FALSE,
  unplayed = FALSE,
  sort = TRUE
)
}
\arguments{
\item{raw}{list of match objects from TBA API as read by tba_readR}

\item{alliances}{(boolean) unpack the alliances?}

\item{breakdown}{(boolean) unpack the score breakdown?}

\item{unplayed}{(boolean) include matches with scores of -1 (indicating that the match has not been played?)}

\item{sort}{(boolean) if TRUE, will sort the output matches by competition level and match number.}
}
\value{
tidy tibble of matches
}
\description{
Takes a list of match objects from TBA's API and returns them as a
tidy tibble.
}
\examples{
tidy_matches(read_event_matches("2016mrcmp"), alliances = TRUE,
    breakdown = TRUE)
}
\author{
Gabriel Krotkov
}
