% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_oprs}
\alias{tidy_oprs}
\title{Tidy OPRs}
\usage{
tidy_oprs(raw)
}
\arguments{
\item{raw}{list of OPR objects}
}
\value{
tidy tibble of event OPRs
}
\description{
Returns a tidy tibble of OPRs given raw lists of OPR objects from TBA API
}
\examples{
tidy_oprs(read_event_oprs("2018mrcmp"))
}
\author{
Gabriel Krotkov
}
