% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_rankings}
\alias{tidy_rankings}
\title{Tidy Rankings}
\usage{
tidy_rankings(raw, trim = TRUE)
}
\arguments{
\item{raw}{(list) rankings list object from TBA API}

\item{unpack}{(bool) unpack the tiebreaker info?}
}
\value{
Tidy tibble of event rankings
}
\description{
Returns a tidy tibble of rankings given raw lists of event rankings
}
\examples{
tidy_rankings(read_event_rankings("2016hop"))
}
\author{
Gabriel Krotkov
}
