% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{tidy_teams}
\alias{tidy_teams}
\title{Tidy Teams}
\usage{
tidy_teams(raw)
}
\arguments{
\item{raw}{TBA API team object list}
}
\value{
tidy tibble of teams
}
\description{
Takes a list of TBA API team objects and converts to a tidy tibble
}
\examples{
tidy_teams(read_event_teams("2022txirv"))
}
\author{
Gabriel Krotkov
}
