% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_locations.R
\name{transform_locations}
\alias{transform_locations}
\title{Transform event locations}
\usage{
transform_locations(data, x, y, dim = c(105, 70), units = "meters")
}
\arguments{
\item{data}{A data frame with event location data.}

\item{x}{A character vector identifying the variables for the x coordinate
locations in \code{data}.}

\item{y}{A character vector identifying the variables for the y coordinate
locations in \code{data}.}

\item{dim}{A numeric vector of length two giving the dimensions of how to
transform the pitch. The x-dimension should be the first component.}

\item{units}{A character vector that provides units for the pitch
dimensions. Use "percent" for when \code{dim = c(100, 100)}.}
}
\value{
A tidy tibble with x and y coordinates transformed according to
\code{dim}. The resulting tibble will have "pitch_dimensions"
and "units" attributes. These will be added or modified depending on if
they exist for \code{data}.
}
\description{
This function transforms the Wyscout event data locations. By default,
the data provided is on a range of 0 - 100\%. Thus, an x-y coordinate pair
of (50, 50) corresponds to midfield.
}
\examples{
# load dplyr for examples
library(dplyr)

# read event data given in package
file_path <- system.file("extdata", "events_england.json", package = "scoutr")
events <- read_events(file_path)

result <- events \%>\%
  select(event_sec:end_y) \%>\%
  transform_locations(x = c("start_x", "end_x"), y = c("start_y", "end_y"))

# verify attributes
attr(result, "pitch_dimensions")
attr(result, "units")

# transform to meters, then transform back
events \%>\%
  select(event_sec:end_y) \%>\%
  transform_locations(x = c("start_x", "end_x"),
                      y = c("start_y", "end_y")) \%>\%
  transform_locations(x = c("start_x", "end_x"),
                      y = c("start_y", "end_y"),
                      dim = c(100, 100), units = "percent")

}
\references{
\emph{Pappalardo, L., Cintia, P., Rossi, A. et al. A public data
set of spatio-temporal match events in soccer competitions. Sci
Data 6, 236 (2019). \url{https://doi.org/10.1038/s41597-019-0247-7}}

All public Wyscout data is available at \url{https://figshare.com/collections/Soccer_match_event_dataset/4415000/2}
}
