% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{transpose_insights}
\alias{transpose_insights}
\title{Transpose Insights}
\usage{
transpose_insights(data)
}
\arguments{
\item{data}{long dataframe of insights}
}
\description{
Helper function for tidy_insights, handles the transpose case.
}
