% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{trim_unplayed}
\alias{trim_unplayed}
\title{Trim Unplayed Matches}
\usage{
trim_unplayed(matches)
}
\arguments{
\item{matches}{dataframe of match objects}
}
\description{
This function takes a dataframe of TBA match objects and trims away any
unplayed matches.
}
\details{
We identify a match as unplayed if both the red and blue alliance
score is -1.
}
