% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{unnest_alliances}
\alias{unnest_alliances}
\title{Unnest Alliances}
\usage{
unnest_alliances(matches)
}
\arguments{
\item{matches}{list of match objects}
}
\value{
match objects with the alliance field broken up into alliance members
}
\description{
Unpacks the columns "blue" and "red" and renames new columns with colors.
Dedicated helper for unpack_alliances, a user shouldn't be calling this.
}
\author{
Gabriel Krotkov
}
