% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{unpack_alliances}
\alias{unpack_alliances}
\title{Unpack Alliances}
\usage{
unpack_alliances(matches)
}
\arguments{
\item{matches}{tibble of matches}
}
\value{
tibble of matches with alliance compositions broken down
}
\description{
Takes a tibble of matches and breaks down the "alliance" column. This is
primarily as a helper for tidy_matches
}
\author{
Gabriel Krotkov
}
