% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_tidyR.r
\name{unpack_breakdown}
\alias{unpack_breakdown}
\title{Unpack Breakdown}
\usage{
unpack_breakdown(matches)
}
\arguments{
\item{matches}{Tibble of TBA match objects (unnested)}
}
\value{
Tibble of matches with additional columns for the score breakdown.
}
\description{
Breaks the "score_breakdown" field into many columns
}
\author{
Gabriel Krotkov
}
