% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w_stepwise}
\alias{w_stepwise}
\title{Stepwise-Optimized OPR Weights}
\format{
An object of class \code{numeric} of length 12.
}
\source{
"Clutch Matches are in the Middle, by Girls of Steel"
}
\usage{
w_stepwise
}
\description{
Linear Regression weights optimized stepwise as outlined by Girls of Steel
in "Clutch Matches are in the Middle"
}
\keyword{datasets}
