% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{week_seed_finish_table}
\alias{week_seed_finish_table}
\title{Event Finish Table by Seed and Week}
\usage{
week_seed_finish_table(wk, year = YEAR, finish = "Winner", size = NULL)
}
\arguments{
\item{wk}{Single integer between 1 and 6 representing the week of competition}

\item{year}{Year of interest, defaults to current year.}

\item{size}{(int) If NULL, include all events. Otherwise, only include
events with the specified number of alliances.}
}
\description{
Given a competition week and year, returns a table showing the distribution
of event seed for a particular finish (default "Winner")
}
\examples{
week_seed_finish_table(1)
week_seed_finish_table(4, 2023)
}
