% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{week_winning_seed_table}
\alias{week_winning_seed_table}
\title{Week Event Wins Table by Seed}
\usage{
week_winning_seed_table(wk, year = YEAR)
}
\arguments{
\item{wk}{Single integer between 1 and 6 representing the week of competition}

\item{year}{Year of interest, defaults to current year.}
}
\description{
Given a competition week and year, returns a table showing the distribution
of event seeds.
}
\examples{
week_winning_seed_table(1)
week_winning_seed_table(4, 2023)
}
