% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.r
\name{weight_rows}
\alias{weight_rows}
\title{Weight Rows}
\usage{
weight_rows(df, w)
}
\arguments{
\item{df}{input dataframe to be weighted}

\item{w}{numeric vector of weights applied to `df`, assumed integers}
}
\description{
Weights the rows of the input dataframe by cutting the dataframe into equal
length bins and multiple-counting the rows for appropriate weight
}
\details{
Weights are applied in uniform length bins (so if weights is
length-2, there will be two bins with 50% of the data each, and if weights
is length-5, there will be five bins with 20% of the data each). The
weighting respects order.
}
