% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{year_events}
\alias{year_events}
\title{Year Events}
\usage{
year_events(year = YEAR, official = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{year}{(int) year of interest}

\item{official}{(logical) get only official events?}

\item{simple}{(bool) simplify event objects?}

\item{keys}{(bool) get keys only?}
}
\value{
tidy tibble of events, or vector if keys = TRUE
}
\description{
Reads events in a given year
}
\examples{
year_events(year = 2022, keys = TRUE)
year_events(year = 2018)
year_events(year = 2014, simple = TRUE)
}
\author{
Gabriel Krotkov
}
