% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{year_opr_swings}
\alias{year_opr_swings}
\title{OPR Swings as events progress}
\usage{
year_opr_swings(year)
}
\arguments{
\item{year}{(int) year of interest}
}
\description{
Compute a dataframe summarizing the perturbation of OPR coefficients as
matches progress over all matches that year.
}
\details{
"mean_diff" in the output data.frame represents the mean change in
standardized OPR coefficients (mean across all competing teams) after the
corresponding match.
}
