% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{year_sb}
\alias{year_sb}
\title{Year (Statbotics)}
\usage{
year_sb(yr)
}
\arguments{
\item{yr}{year of interest, as a single-length int}
}
\description{
Returns a list representing statbotics info for a given year
}
