% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{years_sb}
\alias{years_sb}
\title{Years (Statbotics)}
\usage{
years_sb(params)
}
\arguments{
\item{params}{Named list of parameters to pass to the statbotics API call.}
}
\description{
Return a batch of year objects
}
