% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{district_rankings}
\alias{district_rankings}
\title{District Rankings}
\usage{
district_rankings(district_key, detail = c("none", "separate", "breakdown"))
}
\arguments{
\item{district_key}{(character) TBA legal district key}

\item{detail}{(character) How far should we break down the event
district point scores?}
}
\value{
tidy tibble of event rankings
}
\description{
Reads district rankings
}
\details{
"none" - return the JSON data embedded as a list in a column
"separate" - separate out the individual events, but no further
"breakdown" - separate the individual events, and also break down their
points results.
}
\examples{
district_rankings("2016mar", detail = "breakdown")
district_rankings("2022fit")
}
\author{
Gabriel Krotkov
}
