% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_interfaceR.r
\name{event_coprs}
\alias{event_coprs}
\title{Get Event cOPRs}
\usage{
event_coprs(key)
}
\arguments{
\item{key}{(string) event key}
}
\description{
Wrapper function for read_event_coprs and tidy_event_coprs
}
\author{
Dr. Holt Oliver
}
