% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{events_sb}
\alias{events_sb}
\title{Events (Statbotics)}
\usage{
events_sb(params)
}
\arguments{
\item{params}{Named list of parameters to pass to the statbotics API call.}
}
\description{
Events (Statbotics)
}
