% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ridge.R
\name{fit_event_pridge}
\alias{fit_event_pridge}
\title{Fit Event Prior Ridge}
\usage{
fit_event_pridge(event_key, n_cores = NULL)
}
\arguments{
\item{event_key}{(char) TBA-legal event key (ex. "2025mdsev")}

\item{n_cores}{(int) number of cores to parallelize over. If NULL, will select (max - 1) cores}
}
\description{
Given an event key, selects an optimal lambda using LOOCV and fits the prior
ridge model using pre-event EPA from statbotics as the prior.
}
\details{
Relies on statbotics API to establish priors
}
\examples{
fit_event_pridge("2025mdsev")
fit_event_pridge("2023new", n_cores = 3)
}
