% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{matches_sb}
\alias{matches_sb}
\title{Matches (Statbotics)}
\usage{
matches_sb(...)
}
\arguments{
\item{...}{Parameters passed to the statbotics /v3/{matches} endpoint}
}
\description{
Returns a batch of statbotics match objects conforming to the given parameters
}
