% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMeta.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{mergeMeta}
\usage{
mergeMetadata(x)
}
\arguments{
\item{x}{list of Seurat objects (could be named or un-named, but a named list is recommended)}
}
\value{
A dataframe of merged metadata
}
\description{
Quick way to merge metadata actoss multiple Seurat objects into one master dataframe for plotting purposes
}
\examples{
# list of seurat objects with names
objects <- c("obj1", "obj2", "obj3", "obj4")
# Vector of sample IDs
sampleIDs <- c("S1", "S2", "S3", "S4")
names(objects) <- sampleIDs
metadata <- mergeMetadata(objects)
}
