% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcThreshold.R
\name{qcThreshold}
\alias{qcThreshold}
\title{qcThreshold}
\usage{
qcThreshold(
  metadata,
  variable,
  threshold,
  logTransform,
  sampleColors,
  figDir,
  width = width,
  height = height
)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{variable}{A valid column name present in your metadata dataframe}

\item{threshold}{A threshold for your metadata category of interest}

\item{logTransform}{Logical, whether or not to log transform the y-axis}

\item{sampleColors}{A color vector for samples}

\item{figDir}{A directory path ending in "/" to hold output files}

\item{width}{Figure width}

\item{height}{Figure height}
}
\value{
A ggplot object
}
\description{
Generates publication quality quality control violins showing thresholds. Requires that you concatenate all
sample metadata across multiple objects into one dataframe. Can be used in conjunction with \code{ggpubr::ggarrange}
to plot panelled figure of all QC metrics.
}
\examples{
# Generate individual plots as well as a combined one
arrangedViolins <- ggarrange(qcViolin(metadata, "pct_reads_in_peaks", logTransform = FALSE, colors, figDir, width = 8, height = 8),
qcViolin(metadata, "atac_peak_region_fragments", logTransform = TRUE, colors, figDir, width = 8, height = 8),
qcViolin(metadata, "TSS.enrichment", logTransform = FALSE, colors, figDir, width = 8, height = 8),
qcViolin(metadata, "nucleosome_signal", logTransform = FALSE, colors, figDir, width = 8, height = 8),
qcViolin(metadata, "nFeature_peaks", logTransform = FALSE, colors, figDir, width = 8, height = 8))
ggsave(paste0(figDir, "ATAC_Violin_Summary.png"), arrangedViolins, width = 16, height = 10)
}
