% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcViolin.R
\name{qcViolin}
\alias{qcViolin}
\title{qcViolin}
\usage{
qcViolin(
  metadata,
  variable,
  logTransform,
  sampleColors,
  figDir,
  width = width,
  height = height
)
}
\arguments{
\item{metadata}{A dataframe containing all metadata of interest for samples}

\item{variable}{A valid column name present in your metadata dataframe}

\item{logTransform}{Logical, whether or not to log transform the y-axis}

\item{sampleColors}{A color vector for samples}

\item{figDir}{A directory path ending in "/" to hold output files}

\item{width}{Figure width}

\item{height}{Figure height}
}
\value{
A ggplot object
}
\description{
Generates publication quality quality control violins.
}
\examples{
# Generate individual plots as well as a combined one
arrangedViolins <- ggarrange(
  qcViolin(
    metadata, 
    "pct_reads_in_peaks", 
    logTransform = FALSE, 
    colors = colors, 
    figDir = figDir, 
    width = 8, 
    height = 8),
  qcViolin(
    metadata, 
    "atac_peak_region_fragments", 
    logTransform = TRUE, 
    colors = colors, 
    figDir = figDir, 
    width = 8, 
    height = 8)
   )

}
