% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutR.R
\name{qual_schedule}
\alias{qual_schedule}
\title{Event Schedule}
\usage{
qual_schedule(event_code)
}
\arguments{
\item{event_code}{TBA-legal event key}
}
\description{
Returns a dataframe with the event's qualification match schedule
}
