% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_events}
\alias{read_events}
\title{Read Events}
\usage{
read_events(
  year = NA,
  official = FALSE,
  trim_parents = FALSE,
  simple = FALSE,
  keys = FALSE
)
}
\arguments{
\item{year}{(int or NA) year of interest or NA for all years}

\item{official}{(bool) restrict to only official events?}

\item{trim_parents}{(bool) remove events with divisions under them?}

\item{simple}{(bool) simple format return?}

\item{keys}{(bool) return keys only?}
}
\value{
List of year events
}
\description{
Returns a list of all events in the given year, or all events in all years.
}
\examples{
read_events(year = 2016, official = TRUE, trim_parents = TRUE)
}
\author{
Gabriel Krotkov
}
