% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_readR.r
\name{read_teams}
\alias{read_teams}
\title{Read Teams}
\usage{
read_teams(page_num, year = FALSE, simple = FALSE, keys = FALSE)
}
\arguments{
\item{page_num}{Page index of interest}

\item{year}{year of interest, or FALSE for all years}

\item{simple}{return simple team objects?}

\item{keys}{return keys only?}
}
\value{
Returns a list of team objects, or vector of keys
}
\description{
Returns a list of team objects paginated in 500s (uses 0 indexing)
}
\examples{
read_teams(1)
read_teams(3, keys = TRUE)
}
\author{
Gabriel Krotkov
}
