% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{teams_sb}
\alias{teams_sb}
\title{Teams (Statbotics)}
\usage{
teams_sb(params)
}
\arguments{
\item{params}{Named list of parameters to pass to the statbotics API call.}
}
\description{
Teams (Statbotics)
}
