% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbotics_readR.R
\name{years_sb}
\alias{years_sb}
\title{Years (Statbotics)}
\usage{
years_sb(...)
}
\arguments{
\item{...}{Parameters to be passed to the statbotics /v3/{years}/ endpoint}
}
\description{
Returns a batch of statbotics event objects conforming to the given parameters
}
